Move extensions tests to separate test class
