+++ network-store-integration-test/src/test/java/com/powsybl/network/store/integration/NetworkStoreIT.java
@@ -9,18 +9,11 @@ package com.powsybl.network.store.integration;
 import com.google.common.collect.ImmutableSet;
 import com.google.common.collect.Iterables;
 import com.powsybl.cgmes.conformity.CgmesConformity1Catalog;
-import com.powsybl.cgmes.conformity.CgmesConformity1ModifiedCatalog;
-import com.powsybl.cgmes.conversion.CgmesImport;
-import com.powsybl.cgmes.extensions.*;
-import com.powsybl.cgmes.model.CgmesMetadataModel;
-import com.powsybl.cgmes.model.CgmesSubset;
 import com.powsybl.commons.PowsyblException;
 import com.powsybl.commons.datasource.ReadOnlyDataSource;
 import com.powsybl.commons.datasource.ResourceDataSource;
 import com.powsybl.commons.datasource.ResourceSet;
-import com.powsybl.commons.extensions.Extension;
 import com.powsybl.commons.report.ReportNode;
-import com.powsybl.computation.local.LocalComputationManager;
 import com.powsybl.entsoe.util.EntsoeArea;
 import com.powsybl.entsoe.util.EntsoeAreaImpl;
 import com.powsybl.entsoe.util.EntsoeGeographicalCode;
@@ -29,16 +22,9 @@ import com.powsybl.iidm.network.extensions.*;
 import com.powsybl.iidm.network.test.*;
 import com.powsybl.math.graph.TraverseResult;
 import com.powsybl.network.store.client.NetworkStoreService;
-import com.powsybl.network.store.client.PreloadingStrategy;
-import com.powsybl.network.store.client.RestClient;
-import com.powsybl.network.store.client.RestClientImpl;
 import com.powsybl.network.store.iidm.impl.ConfiguredBusImpl;
 import com.powsybl.network.store.iidm.impl.NetworkFactoryImpl;
 import com.powsybl.network.store.iidm.impl.NetworkImpl;
-import com.powsybl.network.store.model.BaseVoltageSourceAttribute;
-import com.powsybl.network.store.model.CgmesMetadataModelAttributes;
-import com.powsybl.network.store.model.CgmesMetadataModelsAttributes;
-import com.powsybl.network.store.model.CimCharacteristicsAttributes;
 import com.powsybl.network.store.server.NetworkStoreApplication;
 import com.powsybl.ucte.converter.UcteImporter;
 import org.apache.commons.collections4.IterableUtils;
@@ -61,6 +47,7 @@ import java.util.stream.Stream;
 import java.util.stream.StreamSupport;
 
 import static com.powsybl.iidm.network.VariantManagerConstants.INITIAL_VARIANT_ID;
+import static com.powsybl.network.store.integration.TestUtils.*;
 import static org.junit.Assert.*;
 import static org.mockito.Mockito.*;
 
@@ -87,25 +74,9 @@ public class NetworkStoreIT {
     @LocalServerPort
     private int randomServerPort;
 
-    private String getBaseUrl() {
-        return "http://localhost:" + randomServerPort + "/";
-    }
-
-    private NetworkStoreService createNetworkStoreService() {
-        return createNetworkStoreService(null);
-    }
-
-    private NetworkStoreService createNetworkStoreService(RestClientMetrics metrics) {
-        RestClient restClient = new RestClientImpl(getBaseUrl());
-        if (metrics != null) {
-            restClient = new TestRestClient(restClient, metrics);
-        }
-        return new NetworkStoreService(restClient, PreloadingStrategy.NONE);
-    }
-
     @Test
     public void test() {
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             // import new network in the store
             assertTrue(service.getNetworkIds().isEmpty());
             Network network = service.importNetwork(new ResourceDataSource("test", new ResourceSet("/", "test.xiidm")));
@@ -141,7 +112,7 @@ public class NetworkStoreIT {
 
     @Test
     public void nodeBreakerTest() {
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Network network = NetworkTest1Factory.create(service.getNetworkFactory(), "1");
             service.flush(network);
 
@@ -232,12 +203,12 @@ public class NetworkStoreIT {
 
     @Test
     public void svcTest() {
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Network network = NetworkStorageTestCaseFactory.create(service.getNetworkFactory());
             service.flush(network);
         }
 
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
 
             Map<UUID, String> networkIds = service.getNetworkIds();
 
@@ -273,7 +244,7 @@ public class NetworkStoreIT {
         }
 
         // reload modified network
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Map<UUID, String> networkIds = service.getNetworkIds();
             Network readNetwork = service.getNetwork(networkIds.keySet().stream().findFirst().get());
 
@@ -293,12 +264,12 @@ public class NetworkStoreIT {
 
     @Test
     public void testSvcRemove() {
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Network network = NetworkStorageTestCaseFactory.create(service.getNetworkFactory());
             service.flush(network);
         }
 
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Map<UUID, String> networkIds = service.getNetworkIds();
             assertEquals(1, networkIds.size());
             Network readNetwork = service.getNetwork(networkIds.keySet().stream().findFirst().get());
@@ -308,7 +279,7 @@ public class NetworkStoreIT {
             service.flush(readNetwork);
         }
 
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Map<UUID, String> networkIds = service.getNetworkIds();
             assertEquals(1, networkIds.size());
             Network readNetwork = service.getNetwork(networkIds.keySet().stream().findFirst().get());
@@ -318,12 +289,12 @@ public class NetworkStoreIT {
 
     @Test
     public void vscConverterStationTest() {
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Network network = NetworkStorageTestCaseFactory.create(service.getNetworkFactory());
             service.flush(network);
         }
 
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
 
             Map<UUID, String> networkIds = service.getNetworkIds();
 
@@ -381,7 +352,7 @@ public class NetworkStoreIT {
         }
 
         // reload modified network
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Map<UUID, String> networkIds = service.getNetworkIds();
             Network readNetwork = service.getNetwork(networkIds.keySet().stream().findFirst().get());
 
@@ -403,12 +374,12 @@ public class NetworkStoreIT {
 
     @Test
     public void testVscConverterRemove() {
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Network network = NetworkStorageTestCaseFactory.create(service.getNetworkFactory());
             service.flush(network);
         }
 
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Map<UUID, String> networkIds = service.getNetworkIds();
             assertEquals(1, networkIds.size());
             Network readNetwork = service.getNetwork(networkIds.keySet().stream().findFirst().get());
@@ -427,7 +398,7 @@ public class NetworkStoreIT {
             service.flush(readNetwork);
         }
 
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Map<UUID, String> networkIds = service.getNetworkIds();
             assertEquals(1, networkIds.size());
             Network readNetwork = service.getNetwork(networkIds.keySet().stream().findFirst().get());
@@ -438,12 +409,12 @@ public class NetworkStoreIT {
 
     @Test
     public void lccConverterStationTest() {
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Network network = NetworkStorageTestCaseFactory.create(service.getNetworkFactory());
             service.flush(network);
         }
 
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
 
             Map<UUID, String> networkIds = service.getNetworkIds();
 
@@ -473,7 +444,7 @@ public class NetworkStoreIT {
         }
 
         // reload modified network
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Map<UUID, String> networkIds = service.getNetworkIds();
             Network readNetwork = service.getNetwork(networkIds.keySet().stream().findFirst().get());
 
@@ -490,12 +461,12 @@ public class NetworkStoreIT {
 
     @Test
     public void testLccConverterRemove() {
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Network network = NetworkStorageTestCaseFactory.create(service.getNetworkFactory());
             service.flush(network);
         }
 
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Map<UUID, String> networkIds = service.getNetworkIds();
             assertEquals(1, networkIds.size());
             Network readNetwork = service.getNetwork(networkIds.keySet().stream().findFirst().get());
@@ -505,7 +476,7 @@ public class NetworkStoreIT {
             service.flush(readNetwork);
         }
 
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Map<UUID, String> networkIds = service.getNetworkIds();
             assertEquals(1, networkIds.size());
             Network readNetwork = service.getNetwork(networkIds.keySet().stream().findFirst().get());
@@ -515,12 +486,12 @@ public class NetworkStoreIT {
 
     @Test
     public void testLineRemove() {
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Network network = NetworkStorageTestCaseFactory.create(service.getNetworkFactory());
             service.flush(network);
         }
 
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Map<UUID, String> networkIds = service.getNetworkIds();
             assertEquals(1, networkIds.size());
             Network readNetwork = service.getNetwork(networkIds.keySet().stream().findFirst().get());
@@ -530,7 +501,7 @@ public class NetworkStoreIT {
             service.flush(readNetwork);
         }
 
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Map<UUID, String> networkIds = service.getNetworkIds();
             assertEquals(1, networkIds.size());
             Network readNetwork = service.getNetwork(networkIds.keySet().stream().findFirst().get());
@@ -540,12 +511,12 @@ public class NetworkStoreIT {
 
     @Test
     public void testLoadRemove() {
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Network network = NetworkStorageTestCaseFactory.create(service.getNetworkFactory());
             service.flush(network);
         }
 
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Map<UUID, String> networkIds = service.getNetworkIds();
             assertEquals(1, networkIds.size());
             Network readNetwork = service.getNetwork(networkIds.keySet().stream().findFirst().get());
@@ -569,7 +540,7 @@ public class NetworkStoreIT {
             service.flush(readNetwork);
         }
 
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Map<UUID, String> networkIds = service.getNetworkIds();
             assertEquals(1, networkIds.size());
             Network readNetwork = service.getNetwork(networkIds.keySet().stream().findFirst().get());
@@ -579,7 +550,7 @@ public class NetworkStoreIT {
             service.flush(readNetwork);
         }
 
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Map<UUID, String> networkIds = service.getNetworkIds();
             assertEquals(1, networkIds.size());
             Network readNetwork = service.getNetwork(networkIds.keySet().stream().findFirst().get());
@@ -589,12 +560,12 @@ public class NetworkStoreIT {
 
     @Test
     public void testBusBarSectionRemove() {
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Network network = NetworkStorageTestCaseFactory.create(service.getNetworkFactory());
             service.flush(network);
         }
 
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Map<UUID, String> networkIds = service.getNetworkIds();
             assertEquals(1, networkIds.size());
             Network readNetwork = service.getNetwork(networkIds.keySet().stream().findFirst().get());
@@ -609,7 +580,7 @@ public class NetworkStoreIT {
             service.flush(readNetwork);
         }
 
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Map<UUID, String> networkIds = service.getNetworkIds();
             assertEquals(1, networkIds.size());
             Network readNetwork = service.getNetwork(networkIds.keySet().stream().findFirst().get());
@@ -619,7 +590,7 @@ public class NetworkStoreIT {
             service.flush(readNetwork);
         }
 
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Map<UUID, String> networkIds = service.getNetworkIds();
             assertEquals(1, networkIds.size());
             Network readNetwork = service.getNetwork(networkIds.keySet().stream().findFirst().get());
@@ -629,12 +600,12 @@ public class NetworkStoreIT {
 
     @Test
     public void testSubstationRemove() {
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Network network = NetworkStorageTestCaseFactory.create(service.getNetworkFactory());
             service.flush(network);
         }
 
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Map<UUID, String> networkIds = service.getNetworkIds();
             assertEquals(1, networkIds.size());
 
@@ -664,7 +635,7 @@ public class NetworkStoreIT {
             service.flush(readNetwork);
         }
 
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Map<UUID, String> networkIds = service.getNetworkIds();
             assertEquals(1, networkIds.size());
             Network readNetwork = service.getNetwork(networkIds.keySet().stream().findFirst().get());
@@ -686,12 +657,12 @@ public class NetworkStoreIT {
 
     @Test
     public void testSubstationUpdate() {
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Network network = NetworkStorageTestCaseFactory.create(service.getNetworkFactory());
             service.flush(network);
         }
 
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Map<UUID, String> networkIds = service.getNetworkIds();
             assertEquals(1, networkIds.size());
 
@@ -712,7 +683,7 @@ public class NetworkStoreIT {
             service.flush(readNetwork);
         }
 
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Map<UUID, String> networkIds = service.getNetworkIds();
             assertEquals(1, networkIds.size());
 
@@ -727,7 +698,7 @@ public class NetworkStoreIT {
 
     @Test
     public void substationTest() {
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Network network = service.createNetwork("test", "test");
 
             NetworkListener mockedListener = mock(DefaultNetworkListener.class);
@@ -770,7 +741,7 @@ public class NetworkStoreIT {
 
     @Test
     public void voltageLevelTest() {
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Network network = service.createNetwork("test", "test");
 
             NetworkListener mockedListener = mock(DefaultNetworkListener.class);
@@ -817,7 +788,7 @@ public class NetworkStoreIT {
 
     @Test
     public void lineTest() {
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Network network = service.createNetwork("test", "test");
 
             NetworkListener mockedListener = mock(DefaultNetworkListener.class);
@@ -900,7 +871,7 @@ public class NetworkStoreIT {
 
     @Test
     public void batteryTest() {
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Network network = service.createNetwork("test", "test");
 
             NetworkListener mockedListener = mock(DefaultNetworkListener.class);
@@ -958,12 +929,12 @@ public class NetworkStoreIT {
 
     @Test
     public void testBatteryRemove() {
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Network network = NetworkStorageTestCaseFactory.create(service.getNetworkFactory());
             service.flush(network);
         }
 
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Map<UUID, String> networkIds = service.getNetworkIds();
             assertEquals(1, networkIds.size());
             Network readNetwork = service.getNetwork(networkIds.keySet().stream().findFirst().get());
@@ -973,7 +944,7 @@ public class NetworkStoreIT {
             service.flush(readNetwork);
         }
 
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Map<UUID, String> networkIds = service.getNetworkIds();
             assertEquals(1, networkIds.size());
             Network readNetwork = service.getNetwork(networkIds.keySet().stream().findFirst().get());
@@ -983,7 +954,7 @@ public class NetworkStoreIT {
 
     @Test
     public void loadTest() {
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Network network = service.createNetwork("test", "test");
 
             NetworkListener mockedListener = mock(DefaultNetworkListener.class);
@@ -1031,12 +1002,12 @@ public class NetworkStoreIT {
 
     @Test
     public void danglingLineTest() {
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Network network = NetworkStorageTestCaseFactory.create(service.getNetworkFactory());
             service.flush(network);
         }
 
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
 
             Map<UUID, String> networkIds = service.getNetworkIds();
 
@@ -1156,7 +1127,7 @@ public class NetworkStoreIT {
         }
 
         // reload modified network
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Map<UUID, String> networkIds = service.getNetworkIds();
             Network readNetwork = service.getNetwork(networkIds.keySet().stream().findFirst().get());
 
@@ -1192,12 +1163,12 @@ public class NetworkStoreIT {
 
     @Test
     public void hvdcLineTest() {
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Network network = NetworkStorageTestCaseFactory.create(service.getNetworkFactory());
             service.flush(network);
         }
 
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
 
             Map<UUID, String> networkIds = service.getNetworkIds();
 
@@ -1233,7 +1204,7 @@ public class NetworkStoreIT {
         }
 
         // reload modified network
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Map<UUID, String> networkIds = service.getNetworkIds();
             Network readNetwork = service.getNetwork(networkIds.keySet().stream().findFirst().get());
 
@@ -1250,12 +1221,12 @@ public class NetworkStoreIT {
 
     @Test
     public void testHvdcLineRemove() {
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Network network = NetworkStorageTestCaseFactory.create(service.getNetworkFactory());
             service.flush(network);
         }
 
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Map<UUID, String> networkIds = service.getNetworkIds();
             assertEquals(1, networkIds.size());
             Network readNetwork = service.getNetwork(networkIds.keySet().stream().findFirst().get());
@@ -1286,7 +1257,7 @@ public class NetworkStoreIT {
             service.flush(readNetwork);
         }
 
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Map<UUID, String> networkIds = service.getNetworkIds();
             assertEquals(1, networkIds.size());
             Network readNetwork = service.getNetwork(networkIds.keySet().stream().findFirst().get());
@@ -1296,7 +1267,7 @@ public class NetworkStoreIT {
             service.flush(readNetwork);
         }
 
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Map<UUID, String> networkIds = service.getNetworkIds();
             assertEquals(1, networkIds.size());
             Network readNetwork = service.getNetwork(networkIds.keySet().stream().findFirst().get());
@@ -1307,12 +1278,12 @@ public class NetworkStoreIT {
 
     @Test
     public void threeWindingsTransformerTest() {
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Network network = NetworkStorageTestCaseFactory.create(service.getNetworkFactory());
             service.flush(network);
         }
 
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
 
             Map<UUID, String> networkIds = service.getNetworkIds();
 
@@ -1384,7 +1355,7 @@ public class NetworkStoreIT {
         }
 
         // reload modified network
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Map<UUID, String> networkIds = service.getNetworkIds();
             Network readNetwork = service.getNetwork(networkIds.keySet().stream().findFirst().get());
 
@@ -1400,12 +1371,12 @@ public class NetworkStoreIT {
 
     @Test
     public void testThreeWindingsTransformerRemove() {
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Network network = NetworkStorageTestCaseFactory.create(service.getNetworkFactory());
             service.flush(network);
         }
 
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Map<UUID, String> networkIds = service.getNetworkIds();
             assertEquals(1, networkIds.size());
             Network readNetwork = service.getNetwork(networkIds.keySet().stream().findFirst().get());
@@ -1415,7 +1386,7 @@ public class NetworkStoreIT {
             service.flush(readNetwork);
         }
 
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Map<UUID, String> networkIds = service.getNetworkIds();
             assertEquals(1, networkIds.size());
             Network readNetwork = service.getNetwork(networkIds.keySet().stream().findFirst().get());
@@ -1425,12 +1396,12 @@ public class NetworkStoreIT {
 
     @Test
     public void twoWindingsTransformerTest() {
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Network network = NetworkStorageTestCaseFactory.create(service.getNetworkFactory());
             service.flush(network);
         }
 
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
 
             Map<UUID, String> networkIds = service.getNetworkIds();
 
@@ -1467,7 +1438,7 @@ public class NetworkStoreIT {
         }
 
         // reload modified network
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Map<UUID, String> networkIds = service.getNetworkIds();
             Network readNetwork = service.getNetwork(networkIds.keySet().stream().findFirst().get());
 
@@ -1511,12 +1482,12 @@ public class NetworkStoreIT {
 
     @Test
     public void testTwoWindingsTransformerRemove() {
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Network network = NetworkStorageTestCaseFactory.create(service.getNetworkFactory());
             service.flush(network);
         }
 
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Map<UUID, String> networkIds = service.getNetworkIds();
             assertEquals(1, networkIds.size());
             Network readNetwork = service.getNetwork(networkIds.keySet().stream().findFirst().get());
@@ -1526,7 +1497,7 @@ public class NetworkStoreIT {
             service.flush(readNetwork);
         }
 
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Map<UUID, String> networkIds = service.getNetworkIds();
             assertEquals(1, networkIds.size());
             Network readNetwork = service.getNetwork(networkIds.keySet().stream().findFirst().get());
@@ -1536,13 +1507,13 @@ public class NetworkStoreIT {
 
     @Test
     public void internalConnectionsFromCgmesTest() {
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             // import new network in the store
             Network network = service.importNetwork(CgmesConformity1Catalog.miniNodeBreaker().dataSource());
             service.flush(network);
         }
 
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
 
             Map<UUID, String> networkIds = service.getNetworkIds();
 
@@ -1571,237 +1542,13 @@ public class NetworkStoreIT {
     }
 
     @Test
-    public void cgmesExtensionsTest() {
-        try (NetworkStoreService service = createNetworkStoreService()) {
-            // import new network in the store
-            Network network = service.importNetwork(CgmesConformity1Catalog.miniNodeBreaker().dataSource());
-            service.flush(network);
-        }
-
-        try (NetworkStoreService service = createNetworkStoreService()) {
-            Map<UUID, String> networkIds = service.getNetworkIds();
-            assertEquals(1, networkIds.size());
-
-            // get extensions by name
-            Network readNetwork = service.getNetwork(networkIds.keySet().stream().findFirst().get());
-            CgmesMetadataModels cgmesMetadataModels = readNetwork.getExtensionByName("cgmesMetadataModels");
-            CimCharacteristics cimCharacteristics = readNetwork.getExtensionByName("cimCharacteristics");
-
-            Optional<CgmesMetadataModel> cgmesSvMetadataModel = cgmesMetadataModels.getModelForSubset(CgmesSubset.STATE_VARIABLES);
-            assertTrue(cgmesSvMetadataModel.isPresent());
-            CgmesMetadataModel cgmesSvMetadata = cgmesSvMetadataModel.get();
-            assertEquals(573, cgmesSvMetadata.getDescription().length());
-            assertTrue(cgmesSvMetadata.getDescription().contains("CGMES Conformity Assessment"));
-            assertEquals(4, cgmesSvMetadata.getVersion());
-            assertEquals("http://A1.de/Planning/ENTSOE/2", cgmesSvMetadata.getModelingAuthoritySet());
-            assertEquals(3, cgmesSvMetadata.getDependentOn().size());
-
-            Optional<CgmesMetadataModel> cgmesSshMetadataModel = cgmesMetadataModels.getModelForSubset(CgmesSubset.STEADY_STATE_HYPOTHESIS);
-            assertTrue(cgmesSshMetadataModel.isPresent());
-            CgmesMetadataModel cgmesSshMetadata = cgmesSshMetadataModel.get();
-            assertEquals(573, cgmesSshMetadata.getDescription().length());
-            assertTrue(cgmesSshMetadata.getDescription().contains("CGMES Conformity Assessment"));
-            assertEquals(4, cgmesSshMetadata.getVersion());
-            assertEquals("http://A1.de/Planning/ENTSOE/2", cgmesSshMetadata.getModelingAuthoritySet());
-            assertEquals(1, cgmesSshMetadata.getDependentOn().size());
-
-            assertEquals(CgmesTopologyKind.NODE_BREAKER, cimCharacteristics.getTopologyKind());
-            assertEquals(16, cimCharacteristics.getCimVersion());
-
-            // get extensions by class
-            cgmesMetadataModels = readNetwork.getExtension(CgmesMetadataModels.class);
-            cimCharacteristics = readNetwork.getExtension(CimCharacteristics.class);
-            cgmesSvMetadataModel = cgmesMetadataModels.getModelForSubset(CgmesSubset.STATE_VARIABLES);
-
-            assertTrue(cgmesSvMetadataModel.isPresent());
-            cgmesSvMetadata = cgmesSvMetadataModel.get();
-            assertEquals(573, cgmesSvMetadata.getDescription().length());
-            assertTrue(cgmesSvMetadata.getDescription().contains("CGMES Conformity Assessment"));
-            assertEquals(4, cgmesSvMetadata.getVersion());
-            assertEquals("http://A1.de/Planning/ENTSOE/2", cgmesSvMetadata.getModelingAuthoritySet());
-            assertEquals(3, cgmesSvMetadata.getDependentOn().size());
-
-            cgmesSshMetadataModel = cgmesMetadataModels.getModelForSubset(CgmesSubset.STEADY_STATE_HYPOTHESIS);
-            assertTrue(cgmesSshMetadataModel.isPresent());
-            cgmesSshMetadata = cgmesSshMetadataModel.get();
-            assertEquals(573, cgmesSshMetadata.getDescription().length());
-            assertTrue(cgmesSshMetadata.getDescription().contains("CGMES Conformity Assessment"));
-            assertEquals(4, cgmesSshMetadata.getVersion());
-            assertEquals("http://A1.de/Planning/ENTSOE/2", cgmesSshMetadata.getModelingAuthoritySet());
-            assertEquals(1, cgmesSshMetadata.getDependentOn().size());
-
-            assertEquals(CgmesTopologyKind.NODE_BREAKER, cimCharacteristics.getTopologyKind());
-            assertEquals(16, cimCharacteristics.getCimVersion());
-
-            // get all extensions
-            Collection<Extension<Network>> cgmesExtensions = readNetwork.getExtensions();
-            Iterator<Extension<Network>> it = cgmesExtensions.iterator();
-            cgmesMetadataModels = (CgmesMetadataModels) it.next();
-            cimCharacteristics = (CimCharacteristics) it.next();
-
-            cgmesSvMetadataModel = cgmesMetadataModels.getModelForSubset(CgmesSubset.STATE_VARIABLES);
-            assertTrue(cgmesSvMetadataModel.isPresent());
-            cgmesSvMetadata = cgmesSvMetadataModel.get();
-            assertEquals(573, cgmesSvMetadata.getDescription().length());
-            assertTrue(cgmesSvMetadata.getDescription().contains("CGMES Conformity Assessment"));
-            assertEquals(4, cgmesSvMetadata.getVersion());
-            assertEquals("http://A1.de/Planning/ENTSOE/2", cgmesSvMetadata.getModelingAuthoritySet());
-            assertEquals(3, cgmesSvMetadata.getDependentOn().size());
-
-            cgmesSshMetadataModel = cgmesMetadataModels.getModelForSubset(CgmesSubset.STEADY_STATE_HYPOTHESIS);
-            assertTrue(cgmesSshMetadataModel.isPresent());
-            cgmesSshMetadata = cgmesSshMetadataModel.get();
-            assertEquals(573, cgmesSshMetadata.getDescription().length());
-            assertTrue(cgmesSshMetadata.getDescription().contains("CGMES Conformity Assessment"));
-            assertEquals(4, cgmesSshMetadata.getVersion());
-            assertEquals("http://A1.de/Planning/ENTSOE/2", cgmesSshMetadata.getModelingAuthoritySet());
-            assertEquals(1, cgmesSshMetadata.getDependentOn().size());
-
-            assertEquals(CgmesTopologyKind.NODE_BREAKER, cimCharacteristics.getTopologyKind());
-            assertEquals(16, cimCharacteristics.getCimVersion());
-
-            // modify extensions
-            CgmesMetadataModelsAttributes cgmesMetadataModelAttributes = CgmesMetadataModelsAttributes.builder()
-                .models(List.of(new CgmesMetadataModelAttributes(CgmesSubset.STATE_VARIABLES,
-                        "svId", "DescriptionSv", 6, "modelingAuthoritySetSv", List.of("profileSv"),
-                        List.of("dependentOnSv"), List.of("supersedesSv")),
-                    new CgmesMetadataModelAttributes(CgmesSubset.STEADY_STATE_HYPOTHESIS,
-                        "sshId", "DescriptionSsh", 7, "modelingAuthoritySetSsh", List.of("profileSsh"),
-                        List.of("dependentOnSsh"), List.of("supersedesSsh"))))
-                .build();
-
-            ((NetworkImpl) readNetwork).getResource().getAttributes().getExtensionAttributes().put(CgmesMetadataModels.NAME, cgmesMetadataModelAttributes);
-
-            CimCharacteristicsAttributes cimCharacteristicsAttributes = CimCharacteristicsAttributes.builder()
-                .cimVersion(5)
-                .cgmesTopologyKind(CgmesTopologyKind.BUS_BRANCH)
-                .build();
-
-            ((NetworkImpl) readNetwork).updateResource(res -> res.getAttributes().setCimCharacteristics(cimCharacteristicsAttributes));
-
-            service.flush(readNetwork);
-        }
-
-        try (NetworkStoreService service = createNetworkStoreService()) {
-            Map<UUID, String> networkIds = service.getNetworkIds();
-            assertEquals(1, networkIds.size());
-
-            Network readNetwork = service.getNetwork(networkIds.keySet().stream().findFirst().get());
-
-            // get extensions by name
-            CgmesMetadataModels cgmesMetadataModels = readNetwork.getExtensionByName("cgmesMetadataModels");
-            Optional<CgmesMetadataModel> cgmesSvMetadataModel = cgmesMetadataModels.getModelForSubset(CgmesSubset.STATE_VARIABLES);
-            assertTrue(cgmesSvMetadataModel.isPresent());
-            CgmesMetadataModel cgmesSvMetadata = cgmesSvMetadataModel.get();
-            assertEquals("svId", cgmesSvMetadata.getId());
-            assertEquals("DescriptionSv", cgmesSvMetadata.getDescription());
-            assertEquals(6, cgmesSvMetadata.getVersion());
-            assertEquals("modelingAuthoritySetSv", cgmesSvMetadata.getModelingAuthoritySet());
-            assertEquals(1, cgmesSvMetadata.getProfiles().size());
-            assertTrue(cgmesSvMetadata.getProfiles().contains("profileSv"));
-            assertEquals(1, cgmesSvMetadata.getDependentOn().size());
-            assertTrue(cgmesSvMetadata.getDependentOn().contains("dependentOnSv"));
-            assertEquals(1, cgmesSvMetadata.getSupersedes().size());
-            assertTrue(cgmesSvMetadata.getSupersedes().contains("supersedesSv"));
-
-            Optional<CgmesMetadataModel> cgmesSshMetadataModel = cgmesMetadataModels.getModelForSubset(CgmesSubset.STEADY_STATE_HYPOTHESIS);
-            assertTrue(cgmesSshMetadataModel.isPresent());
-            CgmesMetadataModel cgmesSshMetadata = cgmesSshMetadataModel.get();
-            assertEquals("sshId", cgmesSshMetadata.getId());
-            assertEquals("DescriptionSsh", cgmesSshMetadata.getDescription());
-            assertEquals(7, cgmesSshMetadata.getVersion());
-            assertEquals("modelingAuthoritySetSsh", cgmesSshMetadata.getModelingAuthoritySet());
-            assertEquals(1, cgmesSshMetadata.getProfiles().size());
-            assertTrue(cgmesSshMetadata.getProfiles().contains("profileSsh"));
-            assertEquals(1, cgmesSshMetadata.getDependentOn().size());
-            assertTrue(cgmesSshMetadata.getDependentOn().contains("dependentOnSsh"));
-            assertEquals(1, cgmesSshMetadata.getSupersedes().size());
-            assertTrue(cgmesSshMetadata.getSupersedes().contains("supersedesSsh"));
-
-            CimCharacteristics cimCharacteristics = readNetwork.getExtensionByName("cimCharacteristics");
-            assertEquals(CgmesTopologyKind.BUS_BRANCH, cimCharacteristics.getTopologyKind());
-            assertEquals(5, cimCharacteristics.getCimVersion());
-
-            // create new extensions
-            readNetwork.newExtension(CgmesMetadataModelsAdder.class)
-                .newModel()
-                .setSubset(CgmesSubset.STATE_VARIABLES)
-                .setId("svId2")
-                .setDescription("DescriptionSv2")
-                .setVersion(9)
-                .setModelingAuthoritySet("modelingAuthoritySetSv2")
-                .addProfile("profileSv2")
-                .addDependentOn("dependentOnSv2")
-                .addSupersedes("supersedesSv2")
-                .add()
-                .newModel()
-                .setSubset(CgmesSubset.STEADY_STATE_HYPOTHESIS)
-                .setId("sshId2")
-                .setDescription("DescriptionSsh2")
-                .setVersion(3)
-                .setModelingAuthoritySet("modelingAuthoritySetSsh2")
-                .addProfile("profileSsh2")
-                .addDependentOn("dependentOnSsh2")
-                .addSupersedes("supersedesSsh2")
-                .add()
-                .add();
-            readNetwork.newExtension(CimCharacteristicsAdder.class)
-                .setTopologyKind(CgmesTopologyKind.NODE_BREAKER)
-                .setCimVersion(6)
-                .add();
-            service.flush(readNetwork);
-        }
-
-        try (NetworkStoreService service = createNetworkStoreService()) {
-            Map<UUID, String> networkIds = service.getNetworkIds();
-            assertEquals(1, networkIds.size());
-
-            Network readNetwork = service.getNetwork(networkIds.keySet().stream().findFirst().get());
-
-            // get extensions by name
-            CgmesMetadataModels cgmesMetadataModels = readNetwork.getExtensionByName("cgmesMetadataModels");
-            Optional<CgmesMetadataModel> cgmesSvMetadataModel = cgmesMetadataModels.getModelForSubset(CgmesSubset.STATE_VARIABLES);
-            assertTrue(cgmesSvMetadataModel.isPresent());
-            CgmesMetadataModel cgmesSvMetadata = cgmesSvMetadataModel.get();
-            assertEquals("svId2", cgmesSvMetadata.getId());
-            assertEquals("DescriptionSv2", cgmesSvMetadata.getDescription());
-            assertEquals(9, cgmesSvMetadata.getVersion());
-            assertEquals("modelingAuthoritySetSv2", cgmesSvMetadata.getModelingAuthoritySet());
-            assertEquals(1, cgmesSvMetadata.getProfiles().size());
-            assertTrue(cgmesSvMetadata.getProfiles().contains("profileSv2"));
-            assertEquals(1, cgmesSvMetadata.getDependentOn().size());
-            assertTrue(cgmesSvMetadata.getDependentOn().contains("dependentOnSv2"));
-            assertEquals(1, cgmesSvMetadata.getSupersedes().size());
-            assertTrue(cgmesSvMetadata.getSupersedes().contains("supersedesSv2"));
-
-            Optional<CgmesMetadataModel> cgmesSshMetadataModel = cgmesMetadataModels.getModelForSubset(CgmesSubset.STEADY_STATE_HYPOTHESIS);
-            assertTrue(cgmesSshMetadataModel.isPresent());
-            CgmesMetadataModel cgmesSshMetadata = cgmesSshMetadataModel.get();
-            assertEquals("sshId2", cgmesSshMetadata.getId());
-            assertEquals("DescriptionSsh2", cgmesSshMetadata.getDescription());
-            assertEquals(3, cgmesSshMetadata.getVersion());
-            assertEquals("modelingAuthoritySetSsh2", cgmesSshMetadata.getModelingAuthoritySet());
-            assertEquals(1, cgmesSshMetadata.getProfiles().size());
-            assertTrue(cgmesSshMetadata.getProfiles().contains("profileSsh2"));
-            assertEquals(1, cgmesSshMetadata.getDependentOn().size());
-            assertTrue(cgmesSshMetadata.getDependentOn().contains("dependentOnSsh2"));
-            assertEquals(1, cgmesSshMetadata.getSupersedes().size());
-            assertTrue(cgmesSshMetadata.getSupersedes().contains("supersedesSsh2"));
-
-            CimCharacteristics cimCharacteristics = readNetwork.getExtensionByName("cimCharacteristics");
-            assertEquals(CgmesTopologyKind.NODE_BREAKER, cimCharacteristics.getTopologyKind());
-            assertEquals(6, cimCharacteristics.getCimVersion());
-        }
-    }
-
-    @Test
     public void aliasesTest() {
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             // import new network in the store
             service.importNetwork(CgmesConformity1Catalog.miniNodeBreaker().dataSource());
         }
 
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Map<UUID, String> networkIds = service.getNetworkIds();
             assertEquals(1, networkIds.size());
 
@@ -1826,7 +1573,7 @@ public class NetworkStoreIT {
             service.flush(readNetwork);
         }
 
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Map<UUID, String> networkIds = service.getNetworkIds();
             assertEquals(1, networkIds.size());
             NetworkImpl readNetwork = (NetworkImpl) service.getNetwork(networkIds.keySet().stream().findFirst().orElseThrow());
@@ -1843,7 +1590,7 @@ public class NetworkStoreIT {
             threeWT.addAlias("alias_with_type", "typeA");
             service.flush(readNetwork);
         }
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Map<UUID, String> networkIds = service.getNetworkIds();
             assertEquals(1, networkIds.size());
             NetworkImpl readNetwork = (NetworkImpl) service.getNetwork(networkIds.keySet().stream().findFirst().orElseThrow());
@@ -1860,7 +1607,7 @@ public class NetworkStoreIT {
             service.flush(readNetwork);
         }
 
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Map<UUID, String> networkIds = service.getNetworkIds();
             assertEquals(1, networkIds.size());
             NetworkImpl readNetwork = (NetworkImpl) service.getNetwork(networkIds.keySet().stream().findFirst().orElseThrow());
@@ -1874,7 +1621,7 @@ public class NetworkStoreIT {
 
     @Test
     public void connectablesTest() {
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Network network = FourSubstationsNodeBreakerFactory.create(service.getNetworkFactory());
             assertEquals(26, network.getConnectableCount());
             assertEquals(26, IterableUtils.size(network.getConnectables()));
@@ -1883,7 +1630,7 @@ public class NetworkStoreIT {
             service.flush(network);
         }
 
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Map<UUID, String> networkIds = service.getNetworkIds();
 
             assertEquals(1, networkIds.size());
@@ -1900,7 +1647,7 @@ public class NetworkStoreIT {
 
     @Test
     public void moreComplexNodeBreakerTest() {
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Network network = FictitiousSwitchFactory.create(service.getNetworkFactory());
             service.flush(network);
         }
@@ -1908,11 +1655,11 @@ public class NetworkStoreIT {
 
     @Test
     public void testPhaseTapChanger() {
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             service.flush(createTapChangerNetwork(service.getNetworkFactory()));
         }
 
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
 
             Map<UUID, String> networkIds = service.getNetworkIds();
 
@@ -2026,7 +1773,7 @@ public class NetworkStoreIT {
         }
 
         // reload modified network
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Map<UUID, String> networkIds = service.getNetworkIds();
             Network readNetwork = service.getNetwork(networkIds.keySet().stream().findFirst().get());
 
@@ -2039,11 +1786,11 @@ public class NetworkStoreIT {
 
     @Test
     public void testGeneratorMinMaxReactiveLimits() {
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             service.flush(createGeneratorNetwork(service.getNetworkFactory(), ReactiveLimitsKind.MIN_MAX));
         }
 
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Map<UUID, String> networkIds = service.getNetworkIds();
             assertEquals(1, networkIds.size());
             Network readNetwork = service.getNetwork(networkIds.keySet().stream().findFirst().get());
@@ -2091,7 +1838,7 @@ public class NetworkStoreIT {
         }
 
         // reload modified network
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Map<UUID, String> networkIds = service.getNetworkIds();
             Network readNetwork = service.getNetwork(networkIds.keySet().stream().findFirst().get());
 
@@ -2119,11 +1866,11 @@ public class NetworkStoreIT {
 
     @Test
     public void testGeneratorCurveReactiveLimits() {
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             service.flush(createGeneratorNetwork(service.getNetworkFactory(), ReactiveLimitsKind.CURVE));
         }
 
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
 
             Map<UUID, String> networkIds = service.getNetworkIds();
             assertEquals(1, networkIds.size());
@@ -2159,12 +1906,12 @@ public class NetworkStoreIT {
 
     @Test
     public void testGeneratorRemove() {
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Network network = createGeneratorNetwork(service.getNetworkFactory(), ReactiveLimitsKind.MIN_MAX);
             service.flush(network);
         }
 
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Map<UUID, String> networkIds = service.getNetworkIds();
             assertEquals(1, networkIds.size());
             Network readNetwork = service.getNetwork(networkIds.keySet().stream().findFirst().get());
@@ -2174,7 +1921,7 @@ public class NetworkStoreIT {
             service.flush(readNetwork);
         }
 
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Map<UUID, String> networkIds = service.getNetworkIds();
             assertEquals(1, networkIds.size());
             Network readNetwork = service.getNetwork(networkIds.keySet().stream().findFirst().get());
@@ -2184,11 +1931,11 @@ public class NetworkStoreIT {
 
     @Test
     public void testBusBreakerNetwork() {
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             service.flush(EurostagTutorialExample1Factory.create(service.getNetworkFactory()));
         }
 
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
 
             Map<UUID, String> networkIds = service.getNetworkIds();
             assertEquals(1, networkIds.size());
@@ -2231,7 +1978,7 @@ public class NetworkStoreIT {
 
     @Test
     public void testComponentCalculationNetwork() {
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Network network = service.createNetwork("test", "test");
             Substation s1 = network.newSubstation()
                 .setId("S1")
@@ -2331,7 +2078,7 @@ public class NetworkStoreIT {
             Network readNetwork = service.getNetwork(networkIds.keySet().stream().findFirst().get());
         }
 
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Map<UUID, String> networkIds = service.getNetworkIds();
             assertEquals(1, networkIds.size());
 
@@ -2351,11 +2098,11 @@ public class NetworkStoreIT {
 
     @Test
     public void testUcteNetwork() {
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             service.flush(loadUcteNetwork(service.getNetworkFactory()));
         }
 
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Map<UUID, String> networkIds = service.getNetworkIds();
             assertEquals(1, networkIds.size());
             Network readNetwork = service.getNetwork(networkIds.keySet().stream().findFirst().get());
@@ -2473,7 +2220,7 @@ public class NetworkStoreIT {
         }
 
         // reload modified network
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Map<UUID, String> networkIds = service.getNetworkIds();
             Network readNetwork = service.getNetwork(networkIds.keySet().stream().findFirst().get());
 
@@ -2496,11 +2243,11 @@ public class NetworkStoreIT {
 
     @Test
     public void testDanglingLineRemove() {
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             service.flush(createRemoveDL(service.getNetworkFactory()));
         }
 
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Map<UUID, String> networkIds = service.getNetworkIds();
             assertEquals(1, networkIds.size());
             Network readNetwork = service.getNetwork(networkIds.keySet().stream().findFirst().get());
@@ -2508,7 +2255,7 @@ public class NetworkStoreIT {
             service.flush(readNetwork);
         }
 
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Map<UUID, String> networkIds = service.getNetworkIds();
             assertEquals(1, networkIds.size());
             Network readNetwork = service.getNetwork(networkIds.keySet().stream().findFirst().get());
@@ -2539,7 +2286,7 @@ public class NetworkStoreIT {
             service.flush(readNetwork);
         }
 
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Map<UUID, String> networkIds = service.getNetworkIds();
             assertEquals(1, networkIds.size());
             Network readNetwork = service.getNetwork(networkIds.keySet().stream().findFirst().get());
@@ -2549,7 +2296,7 @@ public class NetworkStoreIT {
             service.flush(readNetwork);
         }
 
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Map<UUID, String> networkIds = service.getNetworkIds();
             assertEquals(1, networkIds.size());
             Network readNetwork = service.getNetwork(networkIds.keySet().stream().findFirst().get());
@@ -2561,12 +2308,12 @@ public class NetworkStoreIT {
     @Test
     public void switchesTest() {
         // create network and save it
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             service.flush(createSwitchesNetwork(service.getNetworkFactory()));
         }
 
         // load saved network and modify a switch state
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Map<UUID, String> networkIds = service.getNetworkIds();
             assertEquals(1, networkIds.size());
 
@@ -2588,7 +2335,7 @@ public class NetworkStoreIT {
         }
 
         // reload modified network
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Map<UUID, String> networkIds = service.getNetworkIds();
             Network readNetwork = service.getNetwork(networkIds.keySet().stream().findFirst().get());
 
@@ -2603,12 +2350,12 @@ public class NetworkStoreIT {
     @Test
     public void testNodeBreakerVoltageLevelRemove() {
         // create network and save it
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             service.flush(createSwitchesNetwork(service.getNetworkFactory()));
         }
 
         // load saved network and modify a switch state
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Map<UUID, String> networkIds = service.getNetworkIds();
             assertEquals(1, networkIds.size());
 
@@ -2635,7 +2382,7 @@ public class NetworkStoreIT {
         }
 
         // reload modified network
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Map<UUID, String> networkIds = service.getNetworkIds();
             Network readNetwork = service.getNetwork(networkIds.keySet().stream().findFirst().get());
 
@@ -2650,7 +2397,7 @@ public class NetworkStoreIT {
 
     @Test
     public void testVoltageLevel() {
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Network network = EurostagTutorialExample1Factory.createWithMultipleConnectedComponents(service.getNetworkFactory());
 
             VoltageLevel vl3 = network.getVoltageLevel("VLHV3");
@@ -2763,12 +2510,12 @@ public class NetworkStoreIT {
 
     @Test
     public void configuredBusTest() {
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Network network = NetworkStorageTestCaseFactory.create(service.getNetworkFactory());
             service.flush(network);
         }
 
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
 
             Map<UUID, String> networkIds = service.getNetworkIds();
 
@@ -2804,7 +2551,7 @@ public class NetworkStoreIT {
         }
 
         // reload modified network
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Map<UUID, String> networkIds = service.getNetworkIds();
             Network readNetwork = service.getNetwork(networkIds.keySet().stream().findFirst().get());
 
@@ -2818,7 +2565,7 @@ public class NetworkStoreIT {
 
     @Test
     public void testConfiguredBus() {
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             // import new network in the store
             Network network = service.importNetwork(CgmesConformity1Catalog.smallBusBranch().dataSource());
 
@@ -2960,12 +2707,12 @@ public class NetworkStoreIT {
 
     @Test
     public void testBusBreakerVoltageLevelRemove() {
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             // import new network in the store
             Network network = service.importNetwork(CgmesConformity1Catalog.smallBusBranch().dataSource());
         }
 
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Map<UUID, String> networkIds = service.getNetworkIds();
             assertEquals(1, networkIds.size());
 
@@ -2991,7 +2738,7 @@ public class NetworkStoreIT {
         }
 
         // reload modified network
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Map<UUID, String> networkIds = service.getNetworkIds();
             Network readNetwork = service.getNetwork(networkIds.keySet().stream().findFirst().get());
 
@@ -3236,207 +2983,13 @@ public class NetworkStoreIT {
         return network;
     }
 
-    private static Substation createSubstation(Network n, String id, String name, Country country) {
-        return n.newSubstation()
-            .setId(id)
-            .setName(name)
-            .setCountry(country)
-            .add();
-    }
-
-    private static VoltageLevel createVoltageLevel(Substation s, String id, String name,
-                                                   TopologyKind topology, double vNom, int nodeCount) {
-        VoltageLevel vl = s.newVoltageLevel()
-            .setId(id)
-            .setName(name)
-            .setTopologyKind(topology)
-            .setNominalV(vNom)
-            .add();
-        return vl;
-    }
-
-    private static void createBusBarSection(VoltageLevel vl, String id, String name, int node, int busbarIndex, int sectionIndex) {
-        BusbarSection bbs = vl.getNodeBreakerView().newBusbarSection()
-            .setId(id)
-            .setName(name)
-            .setNode(node)
-            .add();
-        bbs.newExtension(BusbarSectionPositionAdder.class).withBusbarIndex(busbarIndex).withSectionIndex(sectionIndex).add();
-        BusbarSectionPosition bbsp = bbs.getExtension(BusbarSectionPosition.class);
-        assertNotNull(bbsp);
-        assertEquals(busbarIndex, bbsp.getBusbarIndex());
-        assertEquals(sectionIndex, bbsp.getSectionIndex());
-        bbsp = bbs.getExtensionByName("position");
-        assertNotNull(bbsp);
-        assertEquals(busbarIndex, bbsp.getBusbarIndex());
-        assertEquals(sectionIndex, bbsp.getSectionIndex());
-    }
-
-    private static void createSwitch(VoltageLevel vl, String id, String name, SwitchKind kind, boolean retained, boolean open, boolean fictitious, int node1, int node2) {
-        vl.getNodeBreakerView().newSwitch()
-            .setId(id)
-            .setName(name)
-            .setKind(kind)
-            .setRetained(retained)
-            .setOpen(open)
-            .setFictitious(fictitious)
-            .setNode1(node1)
-            .setNode2(node2)
-            .add();
-    }
-
-    private static void createLoad(VoltageLevel vl, String id, String name, String feederName, int feederOrder,
-                                   ConnectablePosition.Direction direction, int node, double p0, double q0) {
-        Load load = vl.newLoad()
-            .setId(id)
-            .setName(name)
-            .setNode(node)
-            .setP0(p0)
-            .setQ0(q0)
-            .add();
-        load.newExtension(ConnectablePositionAdder.class).newFeeder()
-            .withName(feederName).withOrder(feederOrder).withDirection(direction).add().add();
-        ConnectablePosition cp = load.getExtension(ConnectablePosition.class);
-        assertNotNull(cp);
-        assertEquals(feederName, cp.getFeeder().getName().orElseThrow());
-        assertEquals(feederOrder, cp.getFeeder().getOrder().orElseThrow().intValue());
-        assertEquals(direction, cp.getFeeder().getDirection());
-        cp = load.getExtensionByName("position");
-        assertNotNull(cp);
-        assertEquals(feederName, cp.getFeeder().getName().orElseThrow());
-        assertEquals(feederOrder, cp.getFeeder().getOrder().orElseThrow().intValue());
-        assertEquals(direction, cp.getFeeder().getDirection());
-    }
-
-    @Test
-    public void testActivePowerControlExtension() {
-        try (NetworkStoreService service = createNetworkStoreService()) {
-            Network network = EurostagTutorialExample1Factory.create(service.getNetworkFactory());
-            Generator gen = network.getGenerator("GEN");
-            gen.newExtension(ActivePowerControlAdder.class)
-                    .withParticipate(true)
-                    .withDroop(6.3f)
-                    .add();
-            service.flush(network);
-        }
-
-        try (NetworkStoreService service = createNetworkStoreService()) {
-            Network network = service.getNetwork(service.getNetworkIds().keySet().iterator().next());
-            Generator gen = network.getGenerator("GEN");
-            ActivePowerControl<Generator> activePowerControl = gen.getExtension(ActivePowerControl.class);
-            assertNotNull(activePowerControl);
-            assertTrue(activePowerControl.isParticipate());
-            assertEquals(6.3f, activePowerControl.getDroop(), 0f);
-            assertNotNull(gen.getExtensionByName("activePowerControl"));
-        }
-    }
-
-    @Test
-    public void testGeneratorStartup() {
-        try (NetworkStoreService service = createNetworkStoreService()) {
-            Network network = EurostagTutorialExample1Factory.create(service.getNetworkFactory());
-            Generator gen = network.getGenerator("GEN");
-            gen.newExtension(GeneratorStartupAdder.class)
-                    .withPlannedActivePowerSetpoint(1.0)
-                    .withStartupCost(2.0)
-                    .withMarginalCost(3.0)
-                    .withPlannedOutageRate(4.0)
-                    .withForcedOutageRate(5.0)
-                    .add();
-            service.flush(network);
-        }
-
-        try (NetworkStoreService service = createNetworkStoreService()) {
-            Network network = service.getNetwork(service.getNetworkIds().keySet().iterator().next());
-            Generator gen = network.getGenerator("GEN");
-            GeneratorStartup generatorStartup = gen.getExtension(GeneratorStartup.class);
-            assertNotNull(generatorStartup);
-            assertEquals(1.0f, generatorStartup.getPlannedActivePowerSetpoint(), 0f);
-            assertEquals(2.0f, generatorStartup.getStartupCost(), 0f);
-            assertEquals(3.0f, generatorStartup.getMarginalCost(), 0f);
-            assertEquals(4.0f, generatorStartup.getPlannedOutageRate(), 0f);
-            assertEquals(5.0f, generatorStartup.getForcedOutageRate(), 0f);
-            assertNotNull(gen.getExtensionByName(GeneratorStartup.NAME));
-        }
-    }
-
-    @Test
-    public void testGeneratorShortCircuit() {
-        try (NetworkStoreService service = createNetworkStoreService()) {
-            Network network = EurostagTutorialExample1Factory.create(service.getNetworkFactory());
-            Generator gen = network.getGenerator("GEN");
-            assertNull(gen.getExtension(GeneratorShortCircuit.class));
-            assertNull(gen.getExtensionByName(GeneratorShortCircuit.NAME));
-            assertTrue(gen.getExtensions().isEmpty());
-            GeneratorShortCircuitAdder circuitAdder = gen.newExtension(GeneratorShortCircuitAdder.class).withDirectTransX(Double.NaN);
-            assertThrows(PowsyblException.class, () -> circuitAdder.add());
-            circuitAdder.withDirectSubtransX(20.)
-                        .withDirectTransX(30.)
-                        .withStepUpTransformerX(50.)
-                        .add();
-            service.flush(network);
-        }
-
-        try (NetworkStoreService service = createNetworkStoreService()) {
-            Network network = service.getNetwork(service.getNetworkIds().keySet().iterator().next());
-            Generator gen = network.getGenerator("GEN");
-            GeneratorShortCircuit generatorShortCircuit = gen.getExtension(GeneratorShortCircuit.class);
-            assertNotNull(generatorShortCircuit);
-            assertEquals(20., generatorShortCircuit.getDirectSubtransX(), 0);
-            assertEquals(30., generatorShortCircuit.getDirectTransX(), 0);
-            assertEquals(50., generatorShortCircuit.getStepUpTransformerX(), 0);
-            assertNotNull(gen.getExtensionByName(GeneratorShortCircuit.NAME));
-            assertEquals(GeneratorShortCircuit.NAME, generatorShortCircuit.getName());
-
-            assertThrows(PowsyblException.class, () -> generatorShortCircuit.setDirectTransX(Double.NaN));
-            generatorShortCircuit.setDirectSubtransX(23.);
-            generatorShortCircuit.setDirectTransX(32.);
-            generatorShortCircuit.setStepUpTransformerX(44.);
-            assertEquals(23., generatorShortCircuit.getDirectSubtransX(), 0);
-            assertEquals(32., generatorShortCircuit.getDirectTransX(), 0);
-            assertEquals(44., generatorShortCircuit.getStepUpTransformerX(), 0);
-        }
-    }
-
-    @Test
-    public void testIdentifiableShortCircuit() {
-        try (NetworkStoreService service = createNetworkStoreService()) {
-            Network network = EurostagTutorialExample1Factory.create(service.getNetworkFactory());
-            VoltageLevel vl = network.getVoltageLevel("VLGEN");
-            assertNull(vl.getExtension(IdentifiableShortCircuit.class));
-            assertNull(vl.getExtensionByName(IdentifiableShortCircuit.NAME));
-            assertTrue(vl.getExtensions().isEmpty());
-            IdentifiableShortCircuitAdder shortCircuitAdder = vl.newExtension(IdentifiableShortCircuitAdder.class);
-            shortCircuitAdder.withIpMin(80.)
-                    .withIpMax(220.)
-                    .add();
-            service.flush(network);
-        }
-
-        try (NetworkStoreService service = createNetworkStoreService()) {
-            Network network = service.getNetwork(service.getNetworkIds().keySet().iterator().next());
-            VoltageLevel vl = network.getVoltageLevel("VLGEN");
-            IdentifiableShortCircuit identifiableShortCircuit = vl.getExtension(IdentifiableShortCircuit.class);
-            assertNotNull(identifiableShortCircuit);
-            assertEquals(80., identifiableShortCircuit.getIpMin(), 0);
-            assertEquals(220., identifiableShortCircuit.getIpMax(), 0);
-            assertNotNull(vl.getExtensionByName(IdentifiableShortCircuit.NAME));
-            assertEquals(IdentifiableShortCircuit.NAME, identifiableShortCircuit.getName());
-
-            identifiableShortCircuit.setIpMin(90.);
-            identifiableShortCircuit.setIpMax(260.);
-            assertEquals(90., identifiableShortCircuit.getIpMin(), 0);
-            assertEquals(260., identifiableShortCircuit.getIpMax(), 0);
-        }
-    }
-
     @Test
     public void testGetIdentifiable() {
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             service.flush(EurostagTutorialExample1Factory.create(service.getNetworkFactory()));
         }
 
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Network network = service.getNetwork(service.getNetworkIds().keySet().iterator().next());
             Identifiable gen = network.getIdentifiable("GEN");
             assertNotNull(gen);
@@ -3453,156 +3006,14 @@ public class NetworkStoreIT {
         }
     }
 
-    private Network createExtensionsNetwork(NetworkFactory networkFactory) {
-        Network network = networkFactory.createNetwork("Extensions network", "test");
-
-        Substation s1 = createSubstation(network, "s1", "s1", Country.FR);
-        VoltageLevel v1 = createVoltageLevel(s1, "v1", "v1", TopologyKind.NODE_BREAKER, 380.0, 20);
-        createBusBarSection(v1, "1.1", "1.1", 0, 1, 1);
-        createLoad(v1, "v1load", "v1load", "v1load", 1, ConnectablePosition.Direction.TOP, 2, 0., 0.);
-
-        VoltageLevel v2 = createVoltageLevel(s1, "v2", "v2", TopologyKind.NODE_BREAKER, 225.0, 20);
-        createBusBarSection(v2, "2.1", "2.1", 0, 1, 1);
-
-        VoltageLevel v3 = createVoltageLevel(s1, "v3", "v3", TopologyKind.NODE_BREAKER, 100.0, 20);
-        createBusBarSection(v2, "3.1", "3.1", 0, 1, 1);
-
-        TwoWindingsTransformer twt2 = s1.newTwoWindingsTransformer().setId("TWT2")
-            .setName("My two windings transformer").setVoltageLevel1("v1").setVoltageLevel2("v2").setNode1(1)
-            .setNode2(1).setR(0.5).setX(4).setG(0).setB(0).setRatedU1(24).setRatedU2(385).setRatedS(100).add();
-        twt2.newExtension(ConnectablePositionAdder.class).newFeeder1().withName("twt2.1").withOrder(2)
-            .withDirection(ConnectablePosition.Direction.TOP).add().newFeeder2().withName("twt2.2").withOrder(2)
-            .withDirection(ConnectablePosition.Direction.TOP).add().add();
-        ConnectablePosition cptwt2 = twt2.getExtension(ConnectablePosition.class);
-        assertEquals("twt2.1", cptwt2.getFeeder1().getName().orElseThrow());
-        assertEquals(2, cptwt2.getFeeder1().getOrder().orElseThrow().intValue());
-        assertEquals(ConnectablePosition.Direction.TOP, cptwt2.getFeeder1().getDirection());
-        assertEquals("twt2.2", cptwt2.getFeeder2().getName().orElseThrow());
-        assertEquals(2, cptwt2.getFeeder2().getOrder().orElseThrow().intValue());
-        assertEquals(ConnectablePosition.Direction.TOP, cptwt2.getFeeder2().getDirection());
-        cptwt2 = twt2.getExtensionByName("position");
-        assertEquals("twt2.1", cptwt2.getFeeder1().getName().orElseThrow());
-        assertEquals(2, cptwt2.getFeeder1().getOrder().orElseThrow().intValue());
-        assertEquals(ConnectablePosition.Direction.TOP, cptwt2.getFeeder1().getDirection());
-        assertEquals("twt2.2", cptwt2.getFeeder2().getName().orElseThrow());
-        assertEquals(2, cptwt2.getFeeder2().getOrder().orElseThrow().intValue());
-        assertEquals(ConnectablePosition.Direction.TOP, cptwt2.getFeeder2().getDirection());
-
-        ThreeWindingsTransformer twt3 = s1.newThreeWindingsTransformer().setId("TWT3")
-            .setName("Three windings transformer 1").setRatedU0(234).newLeg1().setVoltageLevel("v1").setNode(1)
-            .setR(45).setX(35).setG(25).setB(15).setRatedU(5).add().newLeg2().setVoltageLevel("v2").setNode(1)
-            .setR(47).setX(37).setG(27).setB(17).setRatedU(7).add().newLeg3().setVoltageLevel("v3").setNode(1)
-            .setR(49).setX(39).setG(29).setB(19).setRatedU(9).add().add();
-        twt3.newExtension(ConnectablePositionAdder.class).newFeeder1().withName("twt3.1").withOrder(3)
-            .withDirection(ConnectablePosition.Direction.BOTTOM).add().newFeeder2().withName("twt3.2").withOrder(3)
-            .withDirection(ConnectablePosition.Direction.BOTTOM).add().newFeeder3().withName("twt3.3").withOrder(3)
-            .withDirection(ConnectablePosition.Direction.BOTTOM).add().add();
-
-        ConnectablePosition cptwt3 = twt3.getExtension(ConnectablePosition.class);
-        assertEquals("twt3.1", cptwt3.getFeeder1().getName().orElseThrow());
-        assertEquals(3, cptwt3.getFeeder1().getOrder().orElseThrow().intValue());
-        assertEquals(ConnectablePosition.Direction.BOTTOM, cptwt3.getFeeder1().getDirection());
-        assertEquals("twt3.2", cptwt3.getFeeder2().getName().orElseThrow());
-        assertEquals(3, cptwt3.getFeeder2().getOrder().orElseThrow().intValue());
-        assertEquals(ConnectablePosition.Direction.BOTTOM, cptwt3.getFeeder2().getDirection());
-        assertEquals("twt3.3", cptwt3.getFeeder3().getName().orElseThrow());
-        assertEquals(3, cptwt3.getFeeder3().getOrder().orElseThrow().intValue());
-        assertEquals(ConnectablePosition.Direction.BOTTOM, cptwt3.getFeeder3().getDirection());
-        cptwt3 = twt3.getExtensionByName("position");
-        assertEquals("twt3.1", cptwt3.getFeeder1().getName().orElseThrow());
-        assertEquals(3, cptwt3.getFeeder1().getOrder().orElseThrow().intValue());
-        assertEquals(ConnectablePosition.Direction.BOTTOM, cptwt3.getFeeder1().getDirection());
-        assertEquals("twt3.2", cptwt3.getFeeder2().getName().orElseThrow());
-        assertEquals(3, cptwt3.getFeeder2().getOrder().orElseThrow().intValue());
-        assertEquals(ConnectablePosition.Direction.BOTTOM, cptwt3.getFeeder2().getDirection());
-        assertEquals("twt3.3", cptwt3.getFeeder3().getName().orElseThrow());
-        assertEquals(3, cptwt3.getFeeder3().getOrder().orElseThrow().intValue());
-        assertEquals(ConnectablePosition.Direction.BOTTOM, cptwt3.getFeeder3().getDirection());
-        return network;
-    }
-
-    @Test
-    public void extensionsTest() {
-        // create network and save it
-        try (NetworkStoreService service = createNetworkStoreService()) {
-            service.flush(createExtensionsNetwork(service.getNetworkFactory()));
-        }
-
-        // load saved network
-        try (NetworkStoreService service = createNetworkStoreService()) {
-            Map<UUID, String> networkIds = service.getNetworkIds();
-            assertEquals(1, networkIds.size());
-
-            Network readNetwork = service.getNetwork(networkIds.keySet().stream().findFirst().get());
-            assertEquals("Extensions network", readNetwork.getId());
-
-            Load load = readNetwork.getLoad("v1load");
-            TwoWindingsTransformer twt2 = readNetwork.getTwoWindingsTransformer("TWT2");
-            ThreeWindingsTransformer twt3 = readNetwork.getThreeWindingsTransformer("TWT3");
-            assertNotNull(load);
-            assertNotNull(twt2);
-            assertNotNull(twt3);
-            ConnectablePosition cpload = load.getExtension(ConnectablePosition.class);
-            assertNotNull(cpload);
-            assertEquals("v1load", cpload.getFeeder().getName().orElseThrow());
-            assertEquals(1, cpload.getFeeder().getOrder().orElseThrow().intValue());
-            assertEquals(ConnectablePosition.Direction.TOP, cpload.getFeeder().getDirection());
-            cpload = load.getExtensionByName("position");
-            assertNotNull(cpload);
-            assertEquals("v1load", cpload.getFeeder().getName().orElseThrow());
-            assertEquals(1, cpload.getFeeder().getOrder().orElseThrow().intValue());
-            assertEquals(ConnectablePosition.Direction.TOP, cpload.getFeeder().getDirection());
-
-            ConnectablePosition cptwt2 = twt2.getExtension(ConnectablePosition.class);
-            assertNotNull(cptwt2);
-            assertEquals("twt2.1", cptwt2.getFeeder1().getName().orElseThrow());
-            assertEquals(2, cptwt2.getFeeder1().getOrder().orElseThrow().intValue());
-            assertEquals(ConnectablePosition.Direction.TOP, cptwt2.getFeeder1().getDirection());
-            assertEquals("twt2.2", cptwt2.getFeeder2().getName().orElseThrow());
-            assertEquals(2, cptwt2.getFeeder2().getOrder().orElseThrow().intValue());
-            assertEquals(ConnectablePosition.Direction.TOP, cptwt2.getFeeder2().getDirection());
-            cptwt2 = twt2.getExtensionByName("position");
-            assertNotNull(cptwt2);
-            assertEquals("twt2.1", cptwt2.getFeeder1().getName().orElseThrow());
-            assertEquals(2, cptwt2.getFeeder1().getOrder().orElseThrow().intValue());
-            assertEquals(ConnectablePosition.Direction.TOP, cptwt2.getFeeder1().getDirection());
-            assertEquals("twt2.2", cptwt2.getFeeder2().getName().orElseThrow());
-            assertEquals(2, cptwt2.getFeeder2().getOrder().orElseThrow().intValue());
-            assertEquals(ConnectablePosition.Direction.TOP, cptwt2.getFeeder2().getDirection());
-
-            ConnectablePosition cptwt3 = twt3.getExtension(ConnectablePosition.class);
-            assertNotNull(cptwt3);
-            assertEquals("twt3.1", cptwt3.getFeeder1().getName().orElseThrow());
-            assertEquals(3, cptwt3.getFeeder1().getOrder().orElseThrow().intValue());
-            assertEquals(ConnectablePosition.Direction.BOTTOM, cptwt3.getFeeder1().getDirection());
-            assertEquals("twt3.2", cptwt3.getFeeder2().getName().orElseThrow());
-            assertEquals(3, cptwt3.getFeeder2().getOrder().orElseThrow().intValue());
-            assertEquals(ConnectablePosition.Direction.BOTTOM, cptwt3.getFeeder2().getDirection());
-            assertEquals("twt3.3", cptwt3.getFeeder3().getName().orElseThrow());
-            assertEquals(3, cptwt3.getFeeder3().getOrder().orElseThrow().intValue());
-            assertEquals(ConnectablePosition.Direction.BOTTOM, cptwt3.getFeeder3().getDirection());
-            cptwt3 = twt3.getExtensionByName("position");
-            assertNotNull(cptwt3);
-            assertEquals("twt3.1", cptwt3.getFeeder1().getName().orElseThrow());
-            assertEquals(3, cptwt3.getFeeder1().getOrder().orElseThrow().intValue());
-            assertEquals(ConnectablePosition.Direction.BOTTOM, cptwt3.getFeeder1().getDirection());
-            assertEquals("twt3.2", cptwt3.getFeeder2().getName().orElseThrow());
-            assertEquals(3, cptwt3.getFeeder2().getOrder().orElseThrow().intValue());
-            assertEquals(ConnectablePosition.Direction.BOTTOM, cptwt3.getFeeder2().getDirection());
-            assertEquals("twt3.3", cptwt3.getFeeder3().getName().orElseThrow());
-            assertEquals(3, cptwt3.getFeeder3().getOrder().orElseThrow().intValue());
-            assertEquals(ConnectablePosition.Direction.BOTTOM, cptwt3.getFeeder3().getDirection());
-        }
-    }
-
     @Test
     public void shuntCompensatorTest() {
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Network network = NetworkStorageTestCaseFactory.create(service.getNetworkFactory());
             service.flush(network);
         }
 
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Map<UUID, String> networkIds = service.getNetworkIds();
             assertEquals(1, networkIds.size());
 
@@ -3670,7 +3081,7 @@ public class NetworkStoreIT {
         }
 
         // reload modified network
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Map<UUID, String> networkIds = service.getNetworkIds();
             Network readNetwork = service.getNetwork(networkIds.keySet().stream().findFirst().get());
 
@@ -3696,12 +3107,12 @@ public class NetworkStoreIT {
 
     @Test
     public void testShuntCompensatorRemove() {
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Network network = NetworkStorageTestCaseFactory.create(service.getNetworkFactory());
             service.flush(network);
         }
 
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Map<UUID, String> networkIds = service.getNetworkIds();
             assertEquals(1, networkIds.size());
             Network readNetwork = service.getNetwork(networkIds.keySet().stream().findFirst().get());
@@ -3711,7 +3122,7 @@ public class NetworkStoreIT {
             service.flush(readNetwork);
         }
 
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Map<UUID, String> networkIds = service.getNetworkIds();
             assertEquals(1, networkIds.size());
             Network readNetwork = service.getNetwork(networkIds.keySet().stream().findFirst().get());
@@ -3721,12 +3132,12 @@ public class NetworkStoreIT {
 
     @Test
     public void getIdentifiableNetworkTest() {
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Network network = EurostagTutorialExample1Factory.create(service.getNetworkFactory());
             service.flush(network);
         }
 
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Map<UUID, String> networkIds = service.getNetworkIds();
             Network network = service.getNetwork(networkIds.keySet().stream().findFirst().orElseThrow(AssertionError::new));
             // network is itself an identifiable
@@ -3735,91 +3146,13 @@ public class NetworkStoreIT {
     }
 
     @Test
-    public void coordinatedReactiveControlTest() {
-        try (NetworkStoreService service = createNetworkStoreService()) {
-            Network network = EurostagTutorialExample1Factory.create(service.getNetworkFactory());
-            Generator gen = network.getGenerator("GEN");
-            assertNull(gen.getExtension(CoordinatedReactiveControl.class));
-            assertNull(gen.getExtensionByName("coordinatedReactiveControl"));
-            assertTrue(gen.getExtensions().isEmpty());
-            gen.newExtension(CoordinatedReactiveControlAdder.class)
-                .withQPercent(50)
-                .add();
-            service.flush(network);
-        }
-
-        try (NetworkStoreService service = createNetworkStoreService()) {
-            Map<UUID, String> networkIds = service.getNetworkIds();
-            Network network = service.getNetwork(networkIds.keySet().stream().findFirst().orElseThrow(AssertionError::new));
-            Generator gen = network.getGenerator("GEN");
-            CoordinatedReactiveControl extension = gen.getExtension(CoordinatedReactiveControl.class);
-            assertNotNull(extension);
-            assertEquals(50, extension.getQPercent(), 0);
-            assertNotNull(gen.getExtensionByName("coordinatedReactiveControl"));
-            assertEquals(1, gen.getExtensions().size());
-        }
-    }
-
-    @Test
-    public void generatorEntsoeCategoryTest() {
-        try (NetworkStoreService service = createNetworkStoreService()) {
-            Network network = EurostagTutorialExample1Factory.create(service.getNetworkFactory());
-            Generator gen = network.getGenerator("GEN");
-            assertNull(gen.getExtension(GeneratorEntsoeCategory.class));
-            assertNull(gen.getExtensionByName("entsoeCategory"));
-            assertTrue(gen.getExtensions().isEmpty());
-            gen.newExtension(GeneratorEntsoeCategoryAdder.class)
-                .withCode(50)
-                .add();
-            service.flush(network);
-        }
-
-        try (NetworkStoreService service = createNetworkStoreService()) {
-            Map<UUID, String> networkIds = service.getNetworkIds();
-            Network network = service.getNetwork(networkIds.keySet().stream().findFirst().orElseThrow(AssertionError::new));
-            Generator gen = network.getGenerator("GEN");
-            GeneratorEntsoeCategory extension = gen.getExtension(GeneratorEntsoeCategory.class);
-            assertNotNull(extension);
-            assertEquals(50, extension.getCode());
-            assertNotNull(gen.getExtensionByName("entsoeCategory"));
-            assertEquals(1, gen.getExtensions().size());
-        }
-    }
-
-    @Test
-    public void voltagePerReactivePowerControlTest() {
-        try (NetworkStoreService service = createNetworkStoreService()) {
-            Network network = SvcTestCaseFactory.create(service.getNetworkFactory());
-            StaticVarCompensator svc2 = network.getStaticVarCompensator("SVC2");
-            assertNull(svc2.getExtension(VoltagePerReactivePowerControl.class));
-            assertNull(svc2.getExtensionByName("voltagePerReactivePowerControl"));
-            assertTrue(svc2.getExtensions().isEmpty());
-            svc2.newExtension(VoltagePerReactivePowerControlAdder.class)
-                .withSlope(0.3)
-                .add();
-            service.flush(network);
-        }
-
-        try (NetworkStoreService service = createNetworkStoreService()) {
-            Map<UUID, String> networkIds = service.getNetworkIds();
-            Network network = service.getNetwork(networkIds.keySet().stream().findFirst().orElseThrow(AssertionError::new));
-            StaticVarCompensator svc2 = network.getStaticVarCompensator("SVC2");
-            VoltagePerReactivePowerControl extension = svc2.getExtension(VoltagePerReactivePowerControl.class);
-            assertNotNull(extension);
-            assertEquals(0.3, extension.getSlope(), 0);
-            assertNotNull(svc2.getExtensionByName("voltagePerReactivePowerControl"));
-            assertEquals(1, svc2.getExtensions().size());
-        }
-    }
-
-    @Test
     public void regulatingShuntTest() {
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Network network = ShuntTestCaseFactory.create(service.getNetworkFactory());
             service.flush(network);
         }
 
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Map<UUID, String> networkIds = service.getNetworkIds();
             Network network = service.getNetwork(networkIds.keySet().stream().findFirst().orElseThrow(AssertionError::new));
             ShuntCompensator sc = network.getShuntCompensator("SHUNT");
@@ -3836,7 +3169,7 @@ public class NetworkStoreIT {
             service.flush(network);
         }
 
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Map<UUID, String> networkIds = service.getNetworkIds();
             Network network = service.getNetwork(networkIds.keySet().stream().findFirst().orElseThrow(AssertionError::new));
             assertEquals(1, network.getShuntCompensatorCount());
@@ -3851,7 +3184,7 @@ public class NetworkStoreIT {
 
     @Test
     public void propertiesTest() {
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Network network = EurostagTutorialExample1Factory.create(service.getNetworkFactory());
             Generator gen = network.getGenerator("GEN");
 
@@ -3865,7 +3198,7 @@ public class NetworkStoreIT {
             service.flush(network);
         }
 
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Map<UUID, String> networkIds = service.getNetworkIds();
             Network network = service.getNetwork(networkIds.keySet().stream().findFirst().orElseThrow(AssertionError::new));
             Generator gen = network.getGenerator("GEN");
@@ -3879,7 +3212,7 @@ public class NetworkStoreIT {
 
     @Test
     public void ratedSTest() {
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Network network = ThreeWindingsTransformerNetworkFactory.create(service.getNetworkFactory());
             ThreeWindingsTransformer twt = network.getThreeWindingsTransformer("3WT");
             assertTrue(Double.isNaN(twt.getLeg1().getRatedS()));
@@ -3888,215 +3221,11 @@ public class NetworkStoreIT {
             service.flush(network);
         }
 
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Map<UUID, String> networkIds = service.getNetworkIds();
             Network network = service.getNetwork(networkIds.keySet().stream().findFirst().orElseThrow(AssertionError::new));
             ThreeWindingsTransformer twt = network.getThreeWindingsTransformer("3WT");
             assertEquals(101, twt.getLeg1().getRatedS(), 0);
-        }
-    }
-
-    @Test
-    public void loadDetailExtensionTest() {
-        try (NetworkStoreService service = createNetworkStoreService()) {
-            Network network = SvcTestCaseFactory.create(service.getNetworkFactory());
-            Load load2 = network.getLoad("L2");
-            assertNull(load2.getExtension(LoadDetail.class));
-            assertNull(load2.getExtensionByName("loadDetail"));
-            assertTrue(load2.getExtensions().isEmpty());
-            load2.newExtension(LoadDetailAdder.class)
-                .withFixedActivePower(5.5f)
-                .withFixedReactivePower(2.5f)
-                .withVariableActivePower(3.2f)
-                .withVariableReactivePower(2.1f)
-                .add();
-            assertNotNull(load2.getExtension(LoadDetail.class));
-            assertNotNull(load2.getExtensionByName("loadDetail"));
-            assertFalse(load2.getExtensions().isEmpty());
-            LoadDetail loadDetail = load2.getExtension(LoadDetail.class);
-            assertEquals(5.5f, loadDetail.getFixedActivePower(), 0.1f);
-            assertEquals(2.5f, loadDetail.getFixedReactivePower(), 0.1f);
-            assertEquals(3.2f, loadDetail.getVariableActivePower(), 0.1f);
-            assertEquals(2.1f, loadDetail.getVariableReactivePower(), 0.1f);
-            service.flush(network);
-            loadDetail.setFixedActivePower(7.5f);
-            loadDetail.setFixedReactivePower(4.5f);
-            loadDetail.setVariableActivePower(5.2f);
-            loadDetail.setVariableReactivePower(4.1f);
-            assertEquals(7.5f, loadDetail.getFixedActivePower(), 0.1f);
-            assertEquals(4.5f, loadDetail.getFixedReactivePower(), 0.1f);
-            assertEquals(5.2f, loadDetail.getVariableActivePower(), 0.1f);
-            assertEquals(4.1f, loadDetail.getVariableReactivePower(), 0.1f);
-        }
-
-        try (NetworkStoreService service = createNetworkStoreService()) {
-            Map<UUID, String> networkIds = service.getNetworkIds();
-            Network network = service.getNetwork(networkIds.keySet().stream().findFirst().orElseThrow(AssertionError::new));
-            Load load2 = network.getLoad("L2");
-            assertNotNull(load2.getExtension(LoadDetail.class));
-            assertNotNull(load2.getExtensionByName("loadDetail"));
-            assertFalse(load2.getExtensions().isEmpty());
-            LoadDetail loadDetail = load2.getExtension(LoadDetail.class);
-            assertEquals(5.5f, loadDetail.getFixedActivePower(), 0.1f);
-            assertEquals(2.5f, loadDetail.getFixedReactivePower(), 0.1f);
-            assertEquals(3.2f, loadDetail.getVariableActivePower(), 0.1f);
-            assertEquals(2.1f, loadDetail.getVariableReactivePower(), 0.1f);
-        }
-    }
-
-    @Test
-    public void slackTerminalExtensionTest() {
-        try (NetworkStoreService service = createNetworkStoreService()) {
-            Network network = SvcTestCaseFactory.create(service.getNetworkFactory());
-            VoltageLevel vl = network.getVoltageLevel("VL1");
-            assertNull(vl.getExtension(SlackTerminal.class));
-            assertNull(vl.getExtensionByName("slackTerminal"));
-            assertTrue(vl.getExtensions().isEmpty());
-            assertThrows(PowsyblException.class, () -> vl.newExtension(SlackTerminalAdder.class)
-                .withTerminal(null)
-                .add());
-            assertNull(vl.getExtension(SlackTerminal.class));
-            assertNull(vl.getExtensionByName("slackTerminal"));
-            assertTrue(vl.getExtensions().isEmpty());
-            Generator generator = network.getGenerator("G1");
-            vl.newExtension(SlackTerminalAdder.class)
-                .withTerminal(generator.getTerminal())
-                .add();
-            assertNotNull(vl.getExtension(SlackTerminal.class));
-            assertNotNull(vl.getExtensionByName("slackTerminal"));
-            assertFalse(vl.getExtensions().isEmpty());
-            assertEquals(vl.getExtension(SlackTerminal.class).getTerminal(), generator.getTerminal());
-            service.flush(network);
-        }
-
-        try (NetworkStoreService service = createNetworkStoreService()) {
-            Map<UUID, String> networkIds = service.getNetworkIds();
-            Network network = service.getNetwork(networkIds.keySet().stream().findFirst().orElseThrow(AssertionError::new));
-            VoltageLevel vl = network.getVoltageLevel("VL1");
-            Generator generator = network.getGenerator("G1");
-            vl.newExtension(SlackTerminalAdder.class)
-                .withTerminal(generator.getTerminal())
-                .add();
-            assertNotNull(vl.getExtension(SlackTerminal.class));
-            assertNotNull(vl.getExtensionByName("slackTerminal"));
-            assertFalse(vl.getExtensions().isEmpty());
-            assertEquals(vl.getExtension(SlackTerminal.class).getTerminal(), generator.getTerminal());
-        }
-    }
-
-    @Test
-    public void threeWindingsTransformerPhaseAngleClockExtensionTest() {
-        try (NetworkStoreService service = createNetworkStoreService()) {
-            Network network = ThreeWindingsTransformerNetworkFactory.create(service.getNetworkFactory());
-            Substation substation = network.getSubstation("SUBSTATION");
-            ThreeWindingsTransformer twt = substation.getThreeWindingsTransformerStream().findFirst().orElseThrow(AssertionError::new);
-            assertNull(twt.getExtension(ThreeWindingsTransformerPhaseAngleClock.class));
-            assertNull(twt.getExtensionByName("threeWindingsTransformerPhaseAngleClock"));
-            assertTrue(twt.getExtensions().isEmpty());
-            twt.newExtension(ThreeWindingsTransformerPhaseAngleClockAdder.class).withPhaseAngleClockLeg2(1).withPhaseAngleClockLeg3(2).add();
-            assertNotNull(twt.getExtension(ThreeWindingsTransformerPhaseAngleClock.class));
-            assertNotNull(twt.getExtensionByName("threeWindingsTransformerPhaseAngleClock"));
-            assertFalse(twt.getExtensions().isEmpty());
-            assertEquals(1, twt.getExtension(ThreeWindingsTransformerPhaseAngleClock.class).getPhaseAngleClockLeg2());
-            assertEquals(2, twt.getExtension(ThreeWindingsTransformerPhaseAngleClock.class).getPhaseAngleClockLeg3());
-            service.flush(network);
-        }
-
-        try (NetworkStoreService service = createNetworkStoreService()) {
-            Map<UUID, String> networkIds = service.getNetworkIds();
-            Network network = service.getNetwork(networkIds.keySet().stream().findFirst().orElseThrow(AssertionError::new));
-            Substation substation = network.getSubstation("SUBSTATION");
-            ThreeWindingsTransformer twt = substation.getThreeWindingsTransformerStream().findFirst().orElseThrow(AssertionError::new);
-            assertNotNull(twt.getExtension(ThreeWindingsTransformerPhaseAngleClock.class));
-            assertNotNull(twt.getExtensionByName("threeWindingsTransformerPhaseAngleClock"));
-            assertFalse(twt.getExtensions().isEmpty());
-            assertEquals(1, twt.getExtension(ThreeWindingsTransformerPhaseAngleClock.class).getPhaseAngleClockLeg2());
-            assertEquals(2, twt.getExtension(ThreeWindingsTransformerPhaseAngleClock.class).getPhaseAngleClockLeg3());
-        }
-    }
-
-    @Test
-    public void twoWindingsTransformerPhaseAngleClockExtensionTest() {
-        try (NetworkStoreService service = createNetworkStoreService()) {
-            Network network = EurostagTutorialExample1Factory.create(service.getNetworkFactory());
-
-            TwoWindingsTransformer twt = network.getTwoWindingsTransformer("NGEN_NHV1");
-            assertNull(twt.getExtension(TwoWindingsTransformerPhaseAngleClock.class));
-            assertNull(twt.getExtensionByName("twoWindingsTransformerPhaseAngleClock"));
-            assertTrue(twt.getExtensions().isEmpty());
-            twt.newExtension(TwoWindingsTransformerPhaseAngleClockAdder.class).withPhaseAngleClock(1).add();
-            assertNotNull(twt.getExtension(TwoWindingsTransformerPhaseAngleClock.class));
-            assertNotNull(twt.getExtensionByName("twoWindingsTransformerPhaseAngleClock"));
-            assertFalse(twt.getExtensions().isEmpty());
-            assertEquals(1, twt.getExtension(TwoWindingsTransformerPhaseAngleClock.class).getPhaseAngleClock());
-            service.flush(network);
-        }
-
-        try (NetworkStoreService service = createNetworkStoreService()) {
-            Map<UUID, String> networkIds = service.getNetworkIds();
-            Network network = service.getNetwork(networkIds.keySet().stream().findFirst().orElseThrow(AssertionError::new));
-            Substation substation = network.getSubstation("SUBSTATION");
-            TwoWindingsTransformer twt = network.getTwoWindingsTransformer("NGEN_NHV1");
-            assertNotNull(twt.getExtension(TwoWindingsTransformerPhaseAngleClock.class));
-            assertNotNull(twt.getExtensionByName("twoWindingsTransformerPhaseAngleClock"));
-            assertFalse(twt.getExtensions().isEmpty());
-            assertEquals(1, twt.getExtension(TwoWindingsTransformerPhaseAngleClock.class).getPhaseAngleClock());
-        }
-    }
-
-    @Test
-    public void svcVoltagePerReactivePowerExtensionTest() {
-        try (NetworkStoreService service = createNetworkStoreService()) {
-            Network network = SvcTestCaseFactory.create(service.getNetworkFactory());
-
-            StaticVarCompensator svc = network.getStaticVarCompensator("SVC2");
-            assertNull(svc.getExtension(VoltagePerReactivePowerControl.class));
-            assertNull(svc.getExtensionByName("voltagePerReactivePowerControl"));
-            assertTrue(svc.getExtensions().isEmpty());
-            svc.newExtension(VoltagePerReactivePowerControlAdder.class).withSlope(1).add();
-            assertNotNull(svc.getExtension(VoltagePerReactivePowerControl.class));
-            assertNotNull(svc.getExtensionByName("voltagePerReactivePowerControl"));
-            assertFalse(svc.getExtensions().isEmpty());
-            assertEquals(1, svc.getExtension(VoltagePerReactivePowerControl.class).getSlope(), 0.001);
-            service.flush(network);
-        }
-
-        try (NetworkStoreService service = createNetworkStoreService()) {
-            Map<UUID, String> networkIds = service.getNetworkIds();
-            Network network = service.getNetwork(networkIds.keySet().stream().findFirst().orElseThrow(AssertionError::new));
-            StaticVarCompensator svc = network.getStaticVarCompensator("SVC2");
-            assertNotNull(svc.getExtension(VoltagePerReactivePowerControl.class));
-            assertNotNull(svc.getExtensionByName("voltagePerReactivePowerControl"));
-            assertFalse(svc.getExtensions().isEmpty());
-            assertEquals(1, svc.getExtension(VoltagePerReactivePowerControl.class).getSlope(), 0.001);
-        }
-    }
-
-    @Test
-    public void coordinatedReactiveControlExtensionTest() {
-        try (NetworkStoreService service = createNetworkStoreService()) {
-            Network network = SvcTestCaseFactory.create(service.getNetworkFactory());
-
-            Generator generator = network.getGenerator("G1");
-            assertNull(generator.getExtension(CoordinatedReactiveControl.class));
-            assertNull(generator.getExtensionByName("coordinatedReactiveControl"));
-            assertTrue(generator.getExtensions().isEmpty());
-            generator.newExtension(CoordinatedReactiveControlAdder.class).withQPercent(1).add();
-            assertNotNull(generator.getExtension(CoordinatedReactiveControl.class));
-            assertNotNull(generator.getExtensionByName("coordinatedReactiveControl"));
-            assertFalse(generator.getExtensions().isEmpty());
-            assertEquals(1, generator.getExtension(CoordinatedReactiveControl.class).getQPercent(), 0.001);
-            service.flush(network);
-        }
-
-        try (NetworkStoreService service = createNetworkStoreService()) {
-            Map<UUID, String> networkIds = service.getNetworkIds();
-            Network network = service.getNetwork(networkIds.keySet().stream().findFirst().orElseThrow(AssertionError::new));
-            Generator generator = network.getGenerator("G1");
-            assertNotNull(generator.getExtension(CoordinatedReactiveControl.class));
-            assertNotNull(generator.getExtensionByName("coordinatedReactiveControl"));
-            assertFalse(generator.getExtensions().isEmpty());
-            assertEquals(1, generator.getExtension(CoordinatedReactiveControl.class).getQPercent(), 0.001);
         }
     }
 
@@ -4138,12 +3267,12 @@ public class NetworkStoreIT {
 
     @Test
     public void activeAndApparentPowerLimitsTest() {
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Network network = NetworkStorageTestCaseFactory.create(service.getNetworkFactory());
             service.flush(network);
         }
 
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
 
             Map<UUID, String> networkIds = service.getNetworkIds();
 
@@ -4252,12 +3381,12 @@ public class NetworkStoreIT {
 
     @Test
     public void activePowerLimitsAdderValidationTest() {
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Network network = NetworkStorageTestCaseFactory.create(service.getNetworkFactory());
             service.flush(network);
         }
 
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
 
             Map<UUID, String> networkIds = service.getNetworkIds();
 
@@ -4288,12 +3417,12 @@ public class NetworkStoreIT {
 
     @Test
     public void apparentPowerLimitsAdderValidationTest() {
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Network network = NetworkStorageTestCaseFactory.create(service.getNetworkFactory());
             service.flush(network);
         }
 
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
 
             Map<UUID, String> networkIds = service.getNetworkIds();
 
@@ -4323,353 +3452,8 @@ public class NetworkStoreIT {
     }
 
     @Test
-    public void batteryActivePowerControlTest() {
-        try (NetworkStoreService service = createNetworkStoreService()) {
-            Network network = NetworkStorageTestCaseFactory.create(service.getNetworkFactory());
-            service.flush(network);
-        }
-
-        try (NetworkStoreService service = createNetworkStoreService()) {
-
-            Map<UUID, String> networkIds = service.getNetworkIds();
-
-            assertEquals(1, networkIds.size());
-
-            Network readNetwork = service.getNetwork(networkIds.keySet().stream().findFirst().get());
-
-            assertEquals("networkTestCase", readNetwork.getId());
-
-            assertEquals(1, readNetwork.getBatteryCount());
-
-            Battery battery = readNetwork.getBattery("battery");
-
-            battery.newExtension(ActivePowerControlAdder.class)
-                    .withParticipate(false)
-                    .withDroop(1.0f)
-                    .add();
-            ActivePowerControl activePowerControl = battery.getExtension(ActivePowerControl.class);
-            assertFalse(activePowerControl.isParticipate());
-            assertEquals(1.0f, activePowerControl.getDroop(), 0.01);
-
-            activePowerControl = battery.getExtensionByName("activePowerControl");
-            assertFalse(activePowerControl.isParticipate());
-            assertEquals(1.0f, activePowerControl.getDroop(), 0.01);
-
-            Collection<Extension<Battery>> extensions = battery.getExtensions();
-            assertEquals(1, extensions.size());
-            activePowerControl = (ActivePowerControl) extensions.iterator().next();
-            assertFalse(activePowerControl.isParticipate());
-            assertEquals(1.0f, activePowerControl.getDroop(), 0.01);
-        }
-    }
-
-    @Test
-    public void lccActivePowerControlTest() {
-        try (NetworkStoreService service = createNetworkStoreService()) {
-            Network network = NetworkStorageTestCaseFactory.create(service.getNetworkFactory());
-            service.flush(network);
-        }
-
-        try (NetworkStoreService service = createNetworkStoreService()) {
-
-            Map<UUID, String> networkIds = service.getNetworkIds();
-
-            assertEquals(1, networkIds.size());
-
-            Network readNetwork = service.getNetwork(networkIds.keySet().stream().findFirst().get());
-
-            assertEquals("networkTestCase", readNetwork.getId());
-
-            LccConverterStation lccConverterStation = readNetwork.getLccConverterStation("LCC2");
-
-            assertThrows(UnsupportedOperationException.class, () -> lccConverterStation.newExtension(ActivePowerControlAdder.class).withParticipate(false).withDroop(1.0f).add())
-                .getMessage().contains("Cannot set ActivePowerControl");
-            assertNull(lccConverterStation.getExtension(ActivePowerControl.class));
-        }
-    }
-
-    @Test
-    public void loadActivePowerControlTest() {
-        try (NetworkStoreService service = createNetworkStoreService()) {
-            Network network = NetworkStorageTestCaseFactory.create(service.getNetworkFactory());
-            service.flush(network);
-        }
-
-        try (NetworkStoreService service = createNetworkStoreService()) {
-
-            Map<UUID, String> networkIds = service.getNetworkIds();
-
-            assertEquals(1, networkIds.size());
-
-            Network readNetwork = service.getNetwork(networkIds.keySet().stream().findFirst().get());
-
-            assertEquals("networkTestCase", readNetwork.getId());
-
-            Load load = readNetwork.getLoad("load1");
-
-            assertThrows(UnsupportedOperationException.class, () -> load.newExtension(ActivePowerControlAdder.class).withParticipate(false).withDroop(1.0f).add())
-                .getMessage().contains("Cannot set ActivePowerControl");
-            assertNull(load.getExtension(ActivePowerControl.class));
-        }
-    }
-
-    @Test
-    public void hvdcAngleDroopActivePowerControlExtensionTest() {
-        try (NetworkStoreService service = createNetworkStoreService()) {
-            Network network = HvdcTestNetwork.createVsc(service.getNetworkFactory());
-            HvdcLine hvdcLine = network.getHvdcLine("L");
-            assertNull(hvdcLine.getExtension(HvdcAngleDroopActivePowerControl.class));
-            assertNull(hvdcLine.getExtensionByName("hvdcAngleDroopActivePowerControl"));
-            assertTrue(hvdcLine.getExtensions().isEmpty());
-            hvdcLine.newExtension(HvdcAngleDroopActivePowerControlAdder.class)
-                .withP0(10.0f)
-                .withDroop(5.0f)
-                .withEnabled(true)
-                .add();
-            assertNotNull(hvdcLine.getExtension(HvdcAngleDroopActivePowerControl.class));
-            assertNotNull(hvdcLine.getExtensionByName("hvdcAngleDroopActivePowerControl"));
-            assertFalse(hvdcLine.getExtensions().isEmpty());
-            HvdcAngleDroopActivePowerControl hvdcAngleDroopActivePowerControl = hvdcLine.getExtension(HvdcAngleDroopActivePowerControl.class);
-            assertEquals(10.0f, hvdcAngleDroopActivePowerControl.getP0(), 0.1f);
-            assertEquals(5.0f, hvdcAngleDroopActivePowerControl.getDroop(), 0.1f);
-            assertTrue(hvdcAngleDroopActivePowerControl.isEnabled());
-
-            service.flush(network);
-
-            hvdcAngleDroopActivePowerControl.setP0(20.0f);
-            hvdcAngleDroopActivePowerControl.setDroop(80.0f);
-            hvdcAngleDroopActivePowerControl.setEnabled(false);
-            assertEquals(20.0f, hvdcAngleDroopActivePowerControl.getP0(), 0.1f);
-            assertEquals(80.0f, hvdcAngleDroopActivePowerControl.getDroop(), 0.1f);
-            assertFalse(hvdcAngleDroopActivePowerControl.isEnabled());
-
-            assertEquals("L", hvdcAngleDroopActivePowerControl.getExtendable().getId());
-            assertThrows(IllegalArgumentException.class, () -> hvdcAngleDroopActivePowerControl.setP0(Float.NaN));
-            assertThrows(IllegalArgumentException.class, () -> hvdcAngleDroopActivePowerControl.setDroop(Float.NaN));
-        }
-
-        try (NetworkStoreService service = createNetworkStoreService()) {
-            Map<UUID, String> networkIds = service.getNetworkIds();
-            Network network = service.getNetwork(networkIds.keySet().stream().findFirst().orElseThrow(AssertionError::new));
-
-            HvdcLine hvdcLine = network.getHvdcLine("L");
-            assertNotNull(hvdcLine.getExtension(HvdcAngleDroopActivePowerControl.class));
-            assertNotNull(hvdcLine.getExtensionByName("hvdcAngleDroopActivePowerControl"));
-            assertFalse(hvdcLine.getExtensions().isEmpty());
-
-            HvdcAngleDroopActivePowerControl hvdcAngleDroopActivePowerControl = hvdcLine.getExtension(HvdcAngleDroopActivePowerControl.class);
-            assertEquals(10.0f, hvdcAngleDroopActivePowerControl.getP0(), 0.1f);
-            assertEquals(5.0f, hvdcAngleDroopActivePowerControl.getDroop(), 0.1f);
-            assertTrue(hvdcAngleDroopActivePowerControl.isEnabled());
-        }
-    }
-
-    @Test
-    public void hvdcOperatorActivePowerRangeExtensionTest() {
-        try (NetworkStoreService service = createNetworkStoreService()) {
-            Network network = HvdcTestNetwork.createVsc(service.getNetworkFactory());
-            HvdcLine hvdcLine = network.getHvdcLine("L");
-            assertNull(hvdcLine.getExtension(HvdcOperatorActivePowerRange.class));
-            assertNull(hvdcLine.getExtensionByName("hvdcOperatorActivePowerRange"));
-            assertTrue(hvdcLine.getExtensions().isEmpty());
-            hvdcLine.newExtension(HvdcOperatorActivePowerRangeAdder.class)
-                .withOprFromCS1toCS2(15.0f)
-                .withOprFromCS2toCS1(8.0f)
-                .add();
-            assertNotNull(hvdcLine.getExtension(HvdcOperatorActivePowerRange.class));
-            assertNotNull(hvdcLine.getExtensionByName("hvdcOperatorActivePowerRange"));
-            assertFalse(hvdcLine.getExtensions().isEmpty());
-            HvdcOperatorActivePowerRange hvdcOperatorActivePowerRange = hvdcLine.getExtension(HvdcOperatorActivePowerRange.class);
-            assertEquals(15.0f, hvdcOperatorActivePowerRange.getOprFromCS1toCS2(), 0.1f);
-            assertEquals(8.0f, hvdcOperatorActivePowerRange.getOprFromCS2toCS1(), 0.1f);
-
-            service.flush(network);
-
-            hvdcOperatorActivePowerRange.setOprFromCS1toCS2(30.0f);
-            hvdcOperatorActivePowerRange.setOprFromCS2toCS1(22.0f);
-            assertEquals(30.0f, hvdcOperatorActivePowerRange.getOprFromCS1toCS2(), 0.1f);
-            assertEquals(22.0f, hvdcOperatorActivePowerRange.getOprFromCS2toCS1(), 0.1f);
-
-            assertEquals("L", hvdcOperatorActivePowerRange.getExtendable().getId());
-            assertThrows(IllegalArgumentException.class, () -> hvdcOperatorActivePowerRange.setOprFromCS1toCS2(-1.0f));
-            assertThrows(IllegalArgumentException.class, () -> hvdcOperatorActivePowerRange.setOprFromCS2toCS1(-2.0f));
-        }
-
-        try (NetworkStoreService service = createNetworkStoreService()) {
-            Map<UUID, String> networkIds = service.getNetworkIds();
-            Network network = service.getNetwork(networkIds.keySet().stream().findFirst().orElseThrow(AssertionError::new));
-
-            HvdcLine hvdcLine = network.getHvdcLine("L");
-            assertNotNull(hvdcLine.getExtension(HvdcOperatorActivePowerRange.class));
-            assertNotNull(hvdcLine.getExtensionByName("hvdcOperatorActivePowerRange"));
-            assertFalse(hvdcLine.getExtensions().isEmpty());
-
-            HvdcOperatorActivePowerRange hvdcOperatorActivePowerRange = hvdcLine.getExtension(HvdcOperatorActivePowerRange.class);
-            assertEquals(15.0f, hvdcOperatorActivePowerRange.getOprFromCS1toCS2(), 0.1f);
-            assertEquals(8.0f, hvdcOperatorActivePowerRange.getOprFromCS2toCS1(), 0.1f);
-        }
-    }
-
-    @Test
-    public void cgmesControlAreaDanglingLineTest() {
-        try (NetworkStoreService service = createNetworkStoreService()) {
-            // import new network in the store
-            Network network = service.importNetwork(CgmesConformity1Catalog.microGridBaseCaseBE().dataSource());
-            CgmesControlAreas cgmesControlAreas = network.getExtension(CgmesControlAreas.class);
-            assertNotNull(cgmesControlAreas);
-            assertEquals(0, cgmesControlAreas.getCgmesControlAreas().size());
-        }
-
-        try (NetworkStoreService service = createNetworkStoreService()) {
-            Map<UUID, String> networkIds = service.getNetworkIds();
-            assertEquals(1, networkIds.size());
-            UUID networkUuid = networkIds.keySet().iterator().next();
-
-            Network network = service.getNetwork(networkUuid);
-            CgmesControlAreas cgmesControlAreas = network.getExtension(CgmesControlAreas.class);
-            assertNotNull(cgmesControlAreas);
-            assertEquals(0, cgmesControlAreas.getCgmesControlAreas().size());
-            CgmesControlArea cgmesControlArea = cgmesControlAreas.newCgmesControlArea()
-                .setId("ca1")
-                .setEnergyIdentificationCodeEic("code")
-                .setNetInterchange(1000)
-                .add();
-            cgmesControlArea.add(network.getGenerator("550ebe0d-f2b2-48c1-991f-cebea43a21aa").getTerminal());
-            cgmesControlArea.add(network.getDanglingLine("a16b4a6c-70b1-4abf-9a9d-bd0fa47f9fe4").getBoundary());
-            assertEquals(1, cgmesControlAreas.getCgmesControlAreas().size());
-            CgmesControlArea ca1 = cgmesControlAreas.getCgmesControlArea("ca1");
-            assertNotNull(ca1);
-
-            service.flush(network);
-        }
-
-        try (NetworkStoreService service = createNetworkStoreService()) {
-            Map<UUID, String> networkIds = service.getNetworkIds();
-            assertEquals(1, networkIds.size());
-            UUID networkUuid = networkIds.keySet().iterator().next();
-
-            Network network = service.getNetwork(networkUuid);
-            CgmesControlAreas cgmesControlAreas = network.getExtension(CgmesControlAreas.class);
-            assertNotNull(cgmesControlAreas);
-            assertEquals(1, cgmesControlAreas.getCgmesControlAreas().size());
-            assertTrue(cgmesControlAreas.containsCgmesControlAreaId("ca1"));
-            CgmesControlArea cgmesControlArea = cgmesControlAreas.getCgmesControlArea("ca1");
-            assertEquals("ca1", cgmesControlArea.getId());
-            assertNull(cgmesControlArea.getName());
-            assertEquals("code", cgmesControlArea.getEnergyIdentificationCodeEIC());
-            assertEquals(1000, cgmesControlArea.getNetInterchange(), 0);
-            assertEquals(1, cgmesControlArea.getTerminals().size());
-            assertEquals(1, cgmesControlArea.getBoundaries().size());
-        }
-    }
-
-    @Test
-    public void baseVoltageMappingTest() {
-        try (NetworkStoreService service = createNetworkStoreService()) {
-            // import new network in the store
-            Network network = service.importNetwork(CgmesConformity1Catalog.microGridBaseCaseBE().dataSource());
-            // FIXME to be removed at next powsybl-network-store upgrade
-            // assertNull(network.getExtension(Object.class));
-            assertNull(network.getExtensionByName(""));
-            BaseVoltageMapping baseVoltageMapping = network.getExtension(BaseVoltageMapping.class);
-            assertNotNull(baseVoltageMapping);
-            assertEquals(7, baseVoltageMapping.getBaseVoltages().size());
-            assertFalse(baseVoltageMapping.isBaseVoltageEmpty());
-            var ht = baseVoltageMapping.getBaseVoltage(400.0);
-            assertNotNull(ht);
-            assertEquals("65dd04e792584b3b912374e35dec032e", ht.getId());
-            assertEquals(Source.BOUNDARY, ht.getSource());
-            assertEquals(400.0, ht.getNominalV(), .1);
-
-            assertFalse(baseVoltageMapping.isBaseVoltageMapped(42.0));
-            assertNull(baseVoltageMapping.getBaseVoltage(42.0));
-            // IGN do not remplace IGM
-            baseVoltageMapping.addBaseVoltage(10.5, "somethingIGM", Source.IGM);
-            assertNotEquals("somethingIGM", baseVoltageMapping.getBaseVoltage(10.5).getId());
-            // BOUNDARY replace IGM
-            baseVoltageMapping.addBaseVoltage(10.5, "something", Source.BOUNDARY);
-            var bvs = BaseVoltageSourceAttribute.builder().id("something").nominalV(10.5).source(Source.BOUNDARY).build();
-            assertEquals(bvs, baseVoltageMapping.getBaseVoltage(10.5));
-            // BOUNDARY do not replace BOUNDARY
-            baseVoltageMapping.addBaseVoltage(10.5, "somethingAgain", Source.BOUNDARY);
-            assertNotEquals("somethingAgain", baseVoltageMapping.getBaseVoltage(10.5).getId());
-            // IGM do not replace BOUNDARY
-            baseVoltageMapping.addBaseVoltage(10.5, "somethingIGM", Source.IGM);
-            assertNotEquals("somethingIGM", baseVoltageMapping.getBaseVoltage(10.5).getId());
-
-            baseVoltageMapping.addBaseVoltage(42.0, "somethingElse", Source.IGM);
-            var ft = baseVoltageMapping.getBaseVoltage(42.0);
-            assertEquals("somethingElse", ft.getId());
-            assertEquals(Source.IGM, ft.getSource());
-
-            var baseVolatge = baseVoltageMapping.baseVoltagesByNominalVoltageMap();
-            assertEquals(baseVoltageMapping.getBaseVoltages().size(), baseVolatge.size());
-            service.flush(network);
-
-        }
-        try (NetworkStoreService service = createNetworkStoreService()) {
-            Map<UUID, String> networkIds = service.getNetworkIds();
-            assertEquals(1, networkIds.size());
-            UUID networkUuid = networkIds.keySet().iterator().next();
-
-            Network network = service.getNetwork(networkUuid);
-            BaseVoltageMapping baseVoltageMapping = network.getExtensionByName("baseVoltageMapping");
-
-            var ft = baseVoltageMapping.getBaseVoltage(42.0);
-            assertNotNull(ft);
-        }
-    }
-
-    @Test
-    public void cgmesControlAreaTieLineTest() {
-        try (NetworkStoreService service = createNetworkStoreService()) {
-            // import new network in the store
-            Properties properties = new Properties();
-            properties.put(CgmesImport.IMPORT_CGM_WITH_SUBNETWORKS, "false");
-            service.importNetwork(CgmesConformity1Catalog.microGridBaseCaseAssembled().dataSource(), null, LocalComputationManager.getDefault(), properties);
-        }
-
-        try (NetworkStoreService service = createNetworkStoreService()) {
-            Map<UUID, String> networkIds = service.getNetworkIds();
-            assertEquals(1, networkIds.size());
-            UUID networkUuid = networkIds.keySet().iterator().next();
-
-            Network network = service.getNetwork(networkUuid);
-            CgmesControlAreas cgmesControlAreas = network.getExtension(CgmesControlAreas.class);
-            assertNotNull(cgmesControlAreas);
-
-            assertNotNull(cgmesControlAreas);
-            assertEquals(0, cgmesControlAreas.getCgmesControlAreas().size());
-            CgmesControlArea cgmesControlArea = cgmesControlAreas.newCgmesControlArea()
-                .setId("ca2")
-                .setEnergyIdentificationCodeEic("code2")
-                .setNetInterchange(800)
-                .add();
-            cgmesControlArea.add(((TieLine) network.getTieLine("b18cd1aa-7808-49b9-a7cf-605eaf07b006 + e8acf6b6-99cb-45ad-b8dc-16c7866a4ddc")).getDanglingLine1().getBoundary());
-            assertEquals(1, cgmesControlAreas.getCgmesControlAreas().size());
-
-            service.flush(network);
-        }
-
-        try (NetworkStoreService service = createNetworkStoreService()) {
-            Map<UUID, String> networkIds = service.getNetworkIds();
-            assertEquals(1, networkIds.size());
-            UUID networkUuid = networkIds.keySet().iterator().next();
-
-            Network network = service.getNetwork(networkUuid);
-            CgmesControlAreas cgmesControlAreas = network.getExtension(CgmesControlAreas.class);
-            assertNotNull(cgmesControlAreas);
-
-            assertEquals(1, cgmesControlAreas.getCgmesControlAreas().size());
-            CgmesControlArea cgmesControlArea = cgmesControlAreas.getCgmesControlArea("ca2");
-            assertNotNull(cgmesControlArea);
-            assertEquals(1, cgmesControlArea.getBoundaries().size());
-        }
-    }
-
-    @Test
     public void testImportWithoutFlush() {
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
 
             ReportNode report = ReportNode.newRootReportNode()
                     .withMessageTemplate("test", "test")
@@ -4688,7 +3472,7 @@ public class NetworkStoreIT {
 
     @Test
     public void testImportWithProperties() {
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
 
             ReportNode report = ReportNode.newRootReportNode()
                     .withMessageTemplate("test", "test")
@@ -4709,7 +3493,7 @@ public class NetworkStoreIT {
 
     @Test
     public void testImportWithReport() {
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
 
             ReportNode report = ReportNode.newRootReportNode()
                     .withMessageTemplate("test", "test")
@@ -4728,47 +3512,16 @@ public class NetworkStoreIT {
     }
 
     @Test
-    public void testRemoteReactivePowerControl() {
-        try (NetworkStoreService service = createNetworkStoreService()) {
-            // import new network in the store
-            Network network = service.importNetwork(CgmesConformity1ModifiedCatalog.microGridBaseCaseBEReactivePowerGen().dataSource());
-            service.flush(network);
-        }
-
-        try (NetworkStoreService service = createNetworkStoreService()) {
-
-            Map<UUID, String> networkIds = service.getNetworkIds();
-
-            assertEquals(1, networkIds.size());
-
-            Network readNetwork = service.getNetwork(networkIds.keySet().stream().findFirst().get());
-            Generator g = readNetwork.getGenerator("3a3b27be-b18b-4385-b557-6735d733baf0");
-            RemoteReactivePowerControl ext = g.getExtension(RemoteReactivePowerControl.class);
-            assertNotNull(ext);
-            assertEquals(115.5, ext.getTargetQ(), 0.0);
-            assertTrue(ext.isEnabled());
-            assertSame(readNetwork.getTwoWindingsTransformer("a708c3bc-465d-4fe7-b6ef-6fa6408a62b0").getTerminal2(), ext.getRegulatingTerminal());
-        }
-
-    }
-
-    private ResourceDataSource getResource(String fileName, String path) {
-        return new ResourceDataSource(FilenameUtils.getBaseName(fileName),
-            new ResourceSet(FilenameUtils.getPath(path),
-                FilenameUtils.getName(fileName)));
-    }
-
-    @Test
     public void testVariants() {
         // import network on initial variant
         UUID networkUuid;
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Network network = EurostagTutorialExample1Factory.createWithMoreGenerators(service.getNetworkFactory());
             networkUuid = service.getNetworkUuid(network);
             service.flush(network);
         }
 
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Network network = service.getNetwork(networkUuid);
             assertNotNull(network);
 
@@ -4832,7 +3585,7 @@ public class NetworkStoreIT {
             service.flush(network);
         }
 
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Network network = service.getNetwork(networkUuid);
 
             Load load = network.getLoad("LOAD");
@@ -4880,7 +3633,7 @@ public class NetworkStoreIT {
             service.flush(network);
         }
 
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Network network = service.getNetwork(networkUuid);
 
             // check LOAD still exists on initial variant
@@ -4894,7 +3647,7 @@ public class NetworkStoreIT {
 
         // assert http client type call (for performance regression testing)
         var metrics = new RestClientMetrics();
-        try (NetworkStoreService service = createNetworkStoreService(metrics)) {
+        try (NetworkStoreService service = createNetworkStoreService(metrics, randomServerPort)) {
             Network network = service.getNetwork(networkUuid);
             assertEquals(1, metrics.oneGetterCallCount);
             assertEquals(0, metrics.allGetterCallCount);
@@ -4914,13 +3667,13 @@ public class NetworkStoreIT {
     @Test
     public void emptyCacheCloneTest() {
         UUID networkUuid;
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Network network = EurostagTutorialExample1Factory.create(service.getNetworkFactory());
             networkUuid = service.getNetworkUuid(network);
             service.flush(network);
         }
 
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Network network = service.getNetwork(networkUuid);
             assertNotNull(network);
 
@@ -4936,11 +3689,11 @@ public class NetworkStoreIT {
 
         // Using NetworkStoreService.cloneVariant
         // For an empty cache and buffer this should be the same as network.getVariantManager().cloneVariant()
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             // clone initial variant to variant "v2" while nothing has been cached or modified
             service.cloneVariant(networkUuid, INITIAL_VARIANT_ID, "v2");
         }
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Network network = service.getNetwork(networkUuid);
             assertNotNull(network);
             network.getVariantManager().setWorkingVariant("v2");
@@ -4953,7 +3706,7 @@ public class NetworkStoreIT {
 
         // Using NetworkStoreService.cloneVariant
         // With things in the buffer and cache, after a flush the clone should work.
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Network network = service.getNetwork(networkUuid);
             assertNotNull(network);
             network.getVariantManager().setWorkingVariant("v2");
@@ -4968,7 +3721,7 @@ public class NetworkStoreIT {
             service.flush(network);
             service.cloneVariant(networkUuid, "v2", "v3");
         }
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Network network = service.getNetwork(networkUuid);
             assertNotNull(network);
             network.getVariantManager().setWorkingVariant("v3");
@@ -4980,13 +3733,13 @@ public class NetworkStoreIT {
         }
 
         // Using NetworkStoreService.cloneVariant, testing overwrite error
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             // clone initial variant over existing
             PowsyblException ex = assertThrows(PowsyblException.class,
                 () -> service.cloneVariant(networkUuid, INITIAL_VARIANT_ID, "v3"));
             assertTrue(ex.getMessage().contains("already exists"));
         }
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Network network = service.getNetwork(networkUuid);
             assertNotNull(network);
             network.getVariantManager().setWorkingVariant("v3");
@@ -4998,11 +3751,11 @@ public class NetworkStoreIT {
         }
 
         // Using NetworkStoreService.cloneVariant, testing maybeOverwrite
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             // clone initial variant over existing with mayOverwrite=true
             service.cloneVariant(networkUuid, INITIAL_VARIANT_ID, "v3", true);
         }
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Network network = service.getNetwork(networkUuid);
             assertNotNull(network);
             network.getVariantManager().setWorkingVariant("v3");
@@ -5014,13 +3767,13 @@ public class NetworkStoreIT {
         }
 
         // Using NetworkStoreService.cloneVariant, testing overwrite initial
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             // clone initial variant over existing with mayOverwrite=true
             PowsyblException ex = assertThrows(PowsyblException.class,
                 () -> service.cloneVariant(networkUuid, "v2", INITIAL_VARIANT_ID, true));
             assertTrue(ex.getMessage().contains("forbidden"));
         }
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Network network = service.getNetwork(networkUuid);
             assertNotNull(network);
             network.getVariantManager().setWorkingVariant("v3");
@@ -5036,13 +3789,13 @@ public class NetworkStoreIT {
     public void testVariantRemove() {
         // import network on initial variant
         UUID networkUuid;
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Network network = EurostagTutorialExample1Factory.create(service.getNetworkFactory());
             networkUuid = service.getNetworkUuid(network);
             service.flush(network);
         }
 
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Network network = service.getNetwork(networkUuid);
 
             // there is only initial variant
@@ -5075,7 +3828,7 @@ public class NetworkStoreIT {
             service.flush(network);
         }
 
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Network network = service.getNetwork(networkUuid);
 
             // check that on variant "v", we still have 666 as p0 for LOAD
@@ -5094,7 +3847,7 @@ public class NetworkStoreIT {
     @Test
     public void testVoltageLevelWithoutSubstation() {
         UUID networkUuid;
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Network network = service.createNetwork("networknosubstation", "test");
             networkUuid = service.getNetworkUuid(network);
             network.newVoltageLevel()
@@ -5126,7 +3879,7 @@ public class NetworkStoreIT {
 
             service.flush(network);
         }
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Network network = service.getNetwork(networkUuid);
             VoltageLevel voltageLevel = network.getVoltageLevel("bbVL");
             assertNotNull(voltageLevel);
@@ -5144,11 +3897,11 @@ public class NetworkStoreIT {
 
     @Test
     public void testNanValues() {
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             service.flush(createGeneratorNetwork(service.getNetworkFactory(), ReactiveLimitsKind.MIN_MAX));
         }
 
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Map<UUID, String> networkIds = service.getNetworkIds();
             assertEquals(1, networkIds.size());
             Network readNetwork = service.getNetwork(networkIds.keySet().stream().findFirst().get());
@@ -5167,7 +3920,7 @@ public class NetworkStoreIT {
         }
 
         // reload modified network
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Map<UUID, String> networkIds = service.getNetworkIds();
             Network readNetwork = service.getNetwork(networkIds.keySet().stream().findFirst().get());
 
@@ -5181,7 +3934,7 @@ public class NetworkStoreIT {
 
     @Test
     public void testNpeWithTemporaryLimits() {
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             var network = EurostagTutorialExample1Factory.create(service.getNetworkFactory());
             var l = network.getLine("NHV1_NHV2_1");
             l.newCurrentLimits1()
@@ -5190,7 +3943,7 @@ public class NetworkStoreIT {
             service.flush(network);
         }
 
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Map<UUID, String> networkIds = service.getNetworkIds();
             assertEquals(1, networkIds.size());
             Network network = service.getNetwork(networkIds.keySet().stream().findFirst().orElseThrow());
@@ -5202,14 +3955,14 @@ public class NetworkStoreIT {
 
     @Test
     public void testIncrementalUpdate() {
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Network network = EurostagTutorialExample1Factory.create(service.getNetworkFactory());
             network.getBusView().getBuses(); // force storing calculated topology and connectivity
             service.flush(network);
         }
 
         var metrics = new RestClientMetrics();
-        try (NetworkStoreService service = createNetworkStoreService(metrics)) {
+        try (NetworkStoreService service = createNetworkStoreService(metrics, randomServerPort)) {
             Map<UUID, String> networkIds = service.getNetworkIds();
             UUID networkUuid = networkIds.keySet().stream().findFirst().orElseThrow();
             Network network = service.getNetwork(networkUuid);
@@ -5242,7 +3995,7 @@ public class NetworkStoreIT {
                     metrics.updatedUrls);
         }
 
-        try (NetworkStoreService service = createNetworkStoreService(metrics)) {
+        try (NetworkStoreService service = createNetworkStoreService(metrics, randomServerPort)) {
             Map<UUID, String> networkIds = service.getNetworkIds();
             UUID networkUuid = networkIds.keySet().stream().findFirst().orElseThrow();
             Network network = service.getNetwork(networkUuid);
@@ -5270,47 +4023,17 @@ public class NetworkStoreIT {
 
     @Test
     public void testFixNpeGetIdentifiable() {
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Network network = EurostagTutorialExample1Factory.create(service.getNetworkFactory());
             service.flush(network);
         }
 
-        try (NetworkStoreService service = createNetworkStoreService()) {
+        try (NetworkStoreService service = createNetworkStoreService(randomServerPort)) {
             Map<UUID, String> networkIds = service.getNetworkIds();
             UUID networkUuid = networkIds.keySet().stream().findFirst().orElseThrow();
             Network network = service.getNetwork(networkUuid);
             TwoWindingsTransformer twt2 = (TwoWindingsTransformer) network.getIdentifiable("NHV2_NLOAD");
             assertEquals(2, twt2.getRatioTapChanger().getHighTapPosition());
-        }
-    }
-
-    @Test
-    public void testNetworkExtension() {
-        String filePath = "/network_test1_cgmes_metadata_models.xml";
-        ReadOnlyDataSource dataSource = getResource(filePath, filePath);
-
-        try (NetworkStoreService service = createNetworkStoreService()) {
-            // import new network in the store with extension cgmesMetadataModels
-            Network network = service.importNetwork(dataSource);
-            service.flush(network);
-        }
-
-        try (NetworkStoreService service = createNetworkStoreService()) {
-            Map<UUID, String> networkIds = service.getNetworkIds();
-            assertEquals(1, networkIds.size());
-
-            // check that the stored network contains the extension cgmesMetadataModels
-            UUID initialNetworkUuid = networkIds.keySet().stream().findFirst().get();
-            Network readNetwork = service.getNetwork(initialNetworkUuid);
-            assertNotNull(readNetwork.getExtensionByName("cgmesMetadataModels"));
-
-            // clone network
-            Network clonedNetwork = service.cloneNetwork(initialNetworkUuid, List.of("InitialState"));
-            UUID clonedNetworkUuid = service.getNetworkUuid(clonedNetwork);
-            Network readClonedNetwork = service.getNetwork(clonedNetworkUuid);
-            assertNotNull(readClonedNetwork.getExtensionByName("cgmesMetadataModels"));
-
-            service.deleteAllNetworks();
         }
     }
 }
